% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSPO.R
\name{addObject}
\alias{addObject}
\title{Add SPO columns with correct object}
\usage{
addObject(long_data, pred_tb)
}
\arguments{
\item{long_data}{\link{data.frame} Dataframe in long format. Should contain a \verb{<classname>_ID} column for every class present in the domain and range specifications of \code{pred_tb}.
The \verb{<classname>_ID} columns contain identifiers for each instance of the class. You can generate these with \code{\link[=identify_nodes]{identify_nodes()}}.}

\item{pred_tb}{A tibble containing columns \code{predicate}, \code{rdfs:domain}, \code{rdfs:range} and \code{varname}. The value of \code{varname} is the name of the column in \code{data} corresponding to each \code{predicate}.
You can use \code{\link[=mapPredicates]{mapPredicates()}} and \code{\link[=predicateTibble]{predicateTibble()}} to generate this tibble.}
}
\value{
\link{data.frame} \code{long_data} with the correct object according to the domain of the predicate. Values are only added where the \code{object} column in \code{long_data} is empty.
}
\description{
Adjust dataframe to have correct object according to domain and range specifications in \code{pred_tb}. Helper function used in \code{\link[=pivotLongerSPO]{pivotLongerSPO()}}.
}
\seealso{
\code{\link[=addSubject]{addSubject()}} for examples
}
