% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSPO.R
\name{addSubject}
\alias{addSubject}
\title{Add SPO columns with correct subject}
\usage{
addSubject(long_data, pred_tb)
}
\arguments{
\item{long_data}{\link{data.frame} Dataframe in long format. Should contain a \verb{<classname>_ID} column for every class present in the domain and range specifications of \code{pred_tb}.
The \verb{<classname>_ID} columns contain identifiers for each instance of the class. You can generate these IDs with \code{\link[=identify_nodes]{identify_nodes()}}.}

\item{pred_tb}{A tibble containing columns \code{predicate}, \code{rdfs:domain}, \code{rdfs:range} and \code{varname}. The value of \code{varname} is the name of the column in \code{data} corresponding to each \code{predicate}.
You can use \code{\link[=mapPredicates]{mapPredicates()}} and \code{\link[=predicateTibble]{predicateTibble()}} to generate this tibble.}
}
\value{
\link{data.frame} \code{long_data} with the correct subject for each row according to the domain of the predicate
}
\description{
Adjust dataframe to have correct subject according to domain and range specifications in \code{pred_tb}. Helper function used in \code{\link[=pivotLongerSPO]{pivotLongerSPO()}}.
}
\examples{
\dontrun{
small_pred_tb <- tibble(
  predicate = c("prov:startedAtTime", "prov:endedAtTime", "rdf:type"),
  `rdfs:domain` = c(":Survey", ":Survey", "owl:Thing"),
  `rdfs:range` = c("xsd:dateTime", "xsd:dateTime", ":Interview"),
  varname = c("starttime", "endtime", "Interview_ID")
)

long_data <- tibble(
  subject = c("1", "1", "1"),
  varname = c("starttime", "endtime", "Interview_ID"),
  object = c("2024-02-27 14:18:19", "2024-02-27 14:41:43", "uuid:fe690b51-ee58-4294-b7b6-9f941b47b741"),
  predicate = c("prov:startedAtTime", "prov:endedAtTime", "rdf:type"),
  `rdfs:domain` = c(":Survey", ":Survey", "owl:Thing"),
  `rdfs:range` = c("xsd:dateTime", "xsd:dateTime", ":Interview")
)

long_data <- addSubject(long_data, pred_tb)

# long_data
# A tibble: 3 × 6
#   subject                                    varname      object                 predicate          rdfs:domain  rdfs:range
#   <chr>                                      <chr>        <chr>                  <chr>              <chr>        <chr>
# 1 uuid:fe690b51-ee58-4294-b7b6-9f941b47b741  starttime    2024-02-27 14:18:19    prov:startedAtTime :Survey      xsd:dateTime
# 2 uuid:fe690b51-ee58-4294-b7b6-9f941b47b741. endtime      2024-02-27 14:41:43    prov:endedAtTime   :Survey      xsd:dateTime

}
}
\seealso{
\code{\link[=identify_nodes]{identify_nodes()}}, \code{\link[=pivotLongerSPO]{pivotLongerSPO()}}
}
