% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextHandling.R
\name{addToContext}
\alias{addToContext}
\title{Add to the default context}
\usage{
addToContext(currentConfig, context)
}
\arguments{
\item{currentConfig}{The current configuration list.}

\item{context}{The new context elements to add.}
}
\value{
The updated configuration list.
}
\description{
This function adds to the already existing \code{defaultContext} by merging new
context elements with the existing ones.
}
\examples{
config <- setConfig(ledger = "test1")
c <- list("f" = "https://ns.flur.ee/ledger#", "ex" = "http://example.org/",
          "schema" = "http://schema.org/")
config <- setContext(config, c)
config <- addToContext(config, list("rdfs" = ""))

}
