% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchHandling.R
\name{branchExists}
\alias{branchExists}
\title{Test whether a Fluree branch exists}
\usage{
branchExists(config, branch, ledger = config$ledger)
}
\arguments{
\item{config}{Fluree configuration created by \code{\link[=setConfig]{setConfig()}}.}

\item{branch}{Branch name.}

\item{ledger}{Ledger name. Defaults to \code{config$ledger}.}
}
\value{
A single logical value.
}
\description{
Test whether a Fluree branch exists
}
