% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startUp.R
\name{checkConfiguration}
\alias{checkConfiguration}
\title{Check the configuration parameters}
\usage{
checkConfiguration(config, isConnecting = FALSE)
}
\arguments{
\item{config}{(\code{list()})\cr
The parameters to configure the Fluree instance with.}

\item{isConnecting}{(\code{logical})\cr
If TRUE the instance is attempting to establish a connection to the host.}
}
\description{
This function performs a check on all the configuration parameters to ensure
validity and that they are used in correct combination.
}
