% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredicateInfo.R
\name{checkDTPRange}
\alias{checkDTPRange}
\title{Get datatype properties that have rdfs:range specified}
\usage{
checkDTPRange(ran_res, rdf_data, vn_res)
}
\arguments{
\item{ran_res}{\link{tibble} Containing columns: \code{predicate} and \code{range}. The result of a SPARQL query to find the ranges of all predicates in a list.}

\item{rdf_data}{\link{rdf} Parsed ontology data}

\item{vn_res}{\link{tibble} Containing two columns: \code{predicate} and \code{varname} (possibly NA).}
}
\value{
\link{tibble} Datatype properties that have (at least) a range specified
}
\description{
Helper function in \code{\link[=getPredicateInfo]{getPredicateInfo()}}
}
