% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredicateInfo.R
\name{checkVNDataProp}
\alias{checkVNDataProp}
\title{Check: Variable name is associated with each datatype property}
\usage{
checkVNDataProp(rdf_data, vn_res)
}
\arguments{
\item{rdf_data}{\link{rdf} Parsed ontology data}

\item{vn_res}{\link{tibble} Containing two columns: \code{predicate} and \code{varname} (possibly NA).}
}
\value{
\link{tibble} Datatype properties in \code{pred_vn} for which no corresponding variable name was specified.
}
\description{
Helper function in \code{\link[=getPredicateInfo]{getPredicateInfo()}}.
}
