% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startUp.R
\name{connect}
\alias{connect}
\title{Connect to the Fluree instance}
\usage{
connect()
}
\description{
This function checks that a connection to the running Fluree instance can be
established. The connection needs to be made before any transactions or
queries can be sent.
}
\examples{
config <- list(
   host = 'localhost', 
   port = 58090, 
   ledger = 'policy-view-age', 
   create = TRUE, 
   privateKey = key_get("privateKey", keyring = "Fluree"))

setConfiguration(config = config)

connect()

}
