% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionUtils.R
\name{convertTxnToWhereDelete}
\alias{convertTxnToWhereDelete}
\title{Convert a flattened transaction to a where-delete format}
\usage{
convertTxnToWhereDelete(flattenedTxn, idAlias)
}
\arguments{
\item{flattenedTxn}{(\code{list})
The structured transaction map.}

\item{idAlias}{(\code{character})
The alias used for the @id field in the context of the Fluree instance.}
}
\value{
(\code{list}) A list with \code{where} and \code{delete} keys.
}
\description{
Converts a structured transaction map into a format used for deletion,
producing \code{where} and \code{delete} clauses.
}
