% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_uri_safe.R
\name{create_uri_safe}
\alias{create_uri_safe}
\title{create_uri_safe}
\usage{
create_uri_safe(text)
}
\arguments{
\item{text}{}
}
\description{
Function to create URI-safe identifiers from text strings.
}
\examples{
create_uri_safe("Example Text with Spaces & Special Characters!")
}
