% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBody.R
\name{createBody}
\alias{createBody}
\title{Create Fluree transaction body from dataframe and context tibble}
\usage{
createBody(ledger, data, context)
}
\arguments{
\item{ledger}{\link{string} The name of the ledger into which to transact the data}

\item{data}{\link{data.frame} Containing columns \code{subject}, \code{predicate}, \code{object}}

\item{context}{\link{data.frame} Containing columns \code{prefix}, \code{IRI}}
}
\value{
A JSON-LD object with keys \verb{@context}, \code{ledger}, \code{insert}
}
\description{
Create the body for a Fluree transaction containing fields \verb{@context}, \code{ledger}, and \code{insert}.
Currently only supports insertion of data.
TODO: expand to accommodate queries
}
