% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createChatObject}
\alias{createChatObject}
\title{createChatObject}
\usage{
createChatObject(
  comments = NULL,
  instant = NULL,
  person = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{comments}{character}

\item{instant}{character}

\item{person}{character}
}
\value{
data.frame
}
\description{
function to create chat object
}
