% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlureeObjectFunctions.R
\name{createCollectionObject}
\alias{createCollectionObject}
\title{createCollectionObject}
\usage{
createCollectionObject(
  partition = NULL,
  shard = NULL,
  specdoc = NULL,
  spec = NULL,
  version = NULL,
  doc = NULL,
  name = NULL,
  entityId = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{partition}{character}

\item{shard}{character}

\item{specdoc}{character}

\item{spec}{character}

\item{version}{character}

\item{doc}{character}

\item{name}{character}
}
\value{
data.frame
}
\description{
function to create collection object
}
