% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlureeObjectFunctions.R
\name{createCommentObject}
\alias{createCommentObject}
\title{createCommentObject}
\usage{
createCommentObject(
  message = NULL,
  person = NULL,
  entityId = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{message}{character}

\item{person}{character}
}
\value{
data.frame
}
\description{
function to create comment object
}
