% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createCtxObject}
\alias{createCtxObject}
\title{createCtxObject}
\usage{
createCtxObject(
  doc = NULL,
  fn = NULL,
  key = NULL,
  name = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{doc}{character}

\item{fn}{character}

\item{key}{character}

\item{name}{character}
}
\value{
data.frame
}
\description{
function to create ctx object
}
