% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{createDb}
\alias{createDb}
\title{createDb}
\usage{
createDb(collectionName)
}
\arguments{
\item{collectionName}{Character. The name of the Fluree database.}
}
\value{
character
}
\description{
Helper function to helper create a Fluree database.
}
