% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlureeObjectFunctions.R
\name{createPersonObject}
\alias{createPersonObject}
\title{createPersonObject}
\usage{
createPersonObject(
  fullname = NULL,
  favartists = NULL,
  favnums = NULL,
  follows = NULL,
  handle = NULL,
  entityId = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{fullname}{character}

\item{favartists}{character}

\item{favnums}{character}

\item{follows}{character}

\item{handle}{character}
}
\value{
data.frame
}
\description{
function to create person object
}
