% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{createPredicate}
\alias{createPredicate}
\title{createPredicate}
\usage{
createPredicate(collectionName, schemaList)
}
\arguments{
\item{collectionName}{Character. The name of the Fluree database.}

\item{schemaList}{Character. The list of predicates to add to the database.}
}
\value{
character
}
\description{
Helper function to helper create add predicates to a Fluree database.
}
\examples{
: createSchema("test", [{
 _id: '_predicate',
  name: '_auth/secretValue',
  type: 'string',
})

}
