% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlureeObjectFunctions.R
\name{createPredicateObject}
\alias{createPredicateObject}
\title{createPredicateObject}
\usage{
createPredicateObject(
  retractduplicates = NULL,
  fulltext = NULL,
  restricttag = NULL,
  txspecdoc = NULL,
  txspec = NULL,
  specdoc = NULL,
  deprecated = NULL,
  encrypted = NULL,
  spec = NULL,
  restrictcollection = NULL,
  nohistory = NULL,
  component = NULL,
  upsert = NULL,
  index = NULL,
  multi = NULL,
  unique = NULL,
  type = NULL,
  doc = NULL,
  name = NULL,
  entityId = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{retractduplicates}{character}

\item{fulltext}{character}

\item{restricttag}{character}

\item{txspecdoc}{character}

\item{txspec}{character}

\item{specdoc}{character}

\item{deprecated}{character}

\item{encrypted}{character}

\item{spec}{character}

\item{restrictcollection}{character}

\item{nohistory}{character}

\item{component}{character}

\item{upsert}{character}

\item{index}{character}

\item{multi}{character}

\item{unique}{character}

\item{type}{character}

\item{doc}{character}

\item{name}{character}
}
\value{
data.frame
}
\description{
function to create predicate object
}
