% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createRoleObject}
\alias{createRoleObject}
\title{createRoleObject}
\usage{
createRoleObject(id, doc, rules)
}
\arguments{
\item{id}{character}

\item{doc}{character}

\item{rules}{character}
}
\value{
data.frame
}
\description{
function to create role object
}
