% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlureeObjectFunctions.R
\name{createRuleObject}
\alias{createRuleObject}
\title{createRuleObject}
\usage{
createRuleObject(id, doc, collection, predicates, fns, ops)
}
\arguments{
\item{id}{character}

\item{doc}{character}

\item{collection}{character}

\item{predicates}{character}

\item{fns}{character}

\item{ops}{character}
}
\value{
data.frame
}
\description{
function to create rule object
}
