% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSchema.R
\name{createSchema}
\alias{createSchema}
\title{createSchema}
\usage{
createSchema(path, schema_list, fluree_link = Sys.getenv("fluree_link"))
}
\arguments{
\item{path}{Character. The path to the Fluree database.}

\item{schema_list}{List. The schema to be created (predicates), accepts output from prepareSchema().}

\item{fluree_link}{Character. The link to the Fluree database.}
}
\value{
character
}
\description{
Helper function to create a schema in Fluree
}
