% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createSettingObject}
\alias{createSettingObject}
\title{createSettingObject}
\usage{
createSettingObject(ledgers, id, language)
}
\arguments{
\item{ledgers}{character}

\item{id}{character}

\item{language}{character}
}
\value{
data.frame
}
\description{
function to create setting object
}
