% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createShardObject}
\alias{createShardObject}
\title{createShardObject}
\usage{
createShardObject(
  mutable = NULL,
  miners = NULL,
  name = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{mutable}{character}

\item{miners}{character}

\item{name}{character}
}
\value{
data.frame
}
\description{
function to create shard object
}
