% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createTagObject}
\alias{createTagObject}
\title{createTagObject}
\usage{
createTagObject(doc = NULL, id = NULL, entityId = NULL, deleteObject = FALSE)
}
\arguments{
\item{doc}{character}

\item{id}{character}
}
\value{
data.frame
}
\description{
function to create tag object
}
