% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createUserObject}
\alias{createUserObject}
\title{createUserObject}
\usage{
createUserObject(
  doc = NULL,
  roles = NULL,
  auth = NULL,
  username = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{doc}{character}

\item{roles}{character}

\item{auth}{character}

\item{username}{character}
}
\value{
data.frame
}
\description{
function to create user object
}
