% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cto_to_jsonld.R
\name{cto_to_jsonld}
\alias{cto_to_jsonld}
\title{cto_to_jsonld}
\usage{
cto_to_jsonld(
  formdef,
  base_uri = "https://https://novapc.surveycto.com/",
  instrument = NULL
)
}
\arguments{
\item{formdef}{A data frame representing the form definition from SurveyCTO}

\item{base_uri}{Base URI for the identities formed from the form definition}

\item{instrument}{Name of the instrument (form) being processed}
}
\description{
Standalone function to convert triples to JSON-LD
}
\examples{
\dontrun{
srv <- "KiA_adaptation_ACTIVE"
kia_adapt <- novaCTO::readCTO(srv)
formdef <- kia_adapt$fromschema$kia_adaptation
cto_to_jsonld(formdef,
  base_uri = glue::glue("https://novapc.surveycto.com/{srv}"),
  instrument = "KiA_adaptation_ACTIVE")
}
}
