% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{delete}
\alias{delete}
\title{Delete Subjects by ID}
\usage{
delete(config, id)
}
\arguments{
\item{id}{(\code{list()})\cr
The subject identifier/identifiers to retract from the Fluree database.}
}
\value{
And instance of a delete transaction (as a list).
}
\description{
Deletes all triples for one or more subjects by routing through the Fluree v4
\verb{/update} endpoint with a wildcard where/delete pattern.
}
\examples{
\dontrun{
# Existing data:
#  [
#    { "@id": "freddy", "name": "Freddy" },
#    { "@id": "alice", "name": "Alice" }
#  ]

deleteInstance <- delete(c("freddy"))
sendTransaction(deleteInstance)

# New data state after txn:
#  [
#    { "@id": "alice", "name": "Alice" }
#  ]
}

}
