% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{deleteDb}
\alias{deleteDb}
\title{deleteDb}
\usage{
deleteDb(collectionName)
}
\arguments{
\item{collectionName}{Character. The name of the Fluree database.}
}
\value{
character
}
\description{
Helper function to helper delete to a Fluree database.
}
