% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandIRIs.R
\name{expandIRIs}
\alias{expandIRIs}
\title{expandIRIs}
\usage{
expandIRIs(x, keepCtx = FALSE)
}
\arguments{
\item{x}{List A list representation of a JSON-LD object. If the list contains an
"@context" item, this function will expand all IRIs in the remainder of the
list to full IRIs, and optionally return the list without the "@context" item.}

\item{keepCtx}{Logical Should the "@context" item be included in the result or not?
Defaults to FALSE.}
}
\value{
If x contained an "@context" item, a list with fully expanded IRIs and
(optionally) no "@context" item; otherwise, x unchanged.
}
\description{
Expands the IRIs in a JSON-LD object list according to its
"@context" values.
}
\note{
Largely written by ChatGPT.
}
