% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_turtle.R
\name{export_turtle}
\alias{export_turtle}
\title{export_turtle}
\usage{
export_turtle(triples, context_df = make_surveycto_centext())
}
\arguments{
\item{context_df}{}
}
\description{
Takes the output of \code{map_cto_to_rdf()} and converts it to Turtle format.
}
\examples{
\dontrun{
fromRDF <- map_cto_to_rdf(formdef, base_uri = "https://novapc.surveycto.com/", instrument = "KiA_adaptation_ACTIVE")
export_turtle(fromRDF, context_df = make_surveycto_centext())
}
}
