% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextHandling.R
\name{findIdAlias}
\alias{findIdAlias}
\title{Find the Alias of the ID Field}
\usage{
findIdAlias(context)
}
\arguments{
\item{context}{(\code{list()})\cr
The default context of the Fluree instance.}
}
\value{
(\code{string})
}
\description{
This function looks up if any alias for '@id' has been defined within the
\code{defaultContext} of the current instance. If none is found the default value:
'@id' is used.
}
