% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionUtils.R
\name{flattenEntity}
\alias{flattenEntity}
\title{Flatten an entity into a structured map}
\usage{
flattenEntity(input, idAlias)
}
\arguments{
\item{input}{(\code{list})\cr
The entity or list of entities to flatten.}

\item{idAlias}{(\code{character})\cr
The alias used for the @id field in the context of the Fluree instance.}
}
\value{
(\code{list})
}
\description{
Processes an entity (or a list of entities) and structures it into a map
where each entity is stored with its ID as a key.
}
