% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionUtils.R
\name{flattenTxn}
\alias{flattenTxn}
\title{Flatten a transaction}
\usage{
flattenTxn(txn, idAlias)
}
\arguments{
\item{txn}{(\code{list})\cr
The transaction to flatten.}

\item{idAlias}{(\code{character})\cr
The alias used for the @id field in the context of the Fluree instance.}
}
\value{
(\code{list})
}
\description{
Calls \code{flattenEntity} to process the transaction into a structured map.
}
