% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{fluree_context}
\alias{fluree_context}
\title{The default JSON-LD context of a connection}
\usage{
fluree_context(con)
}
\arguments{
\item{con}{A \code{fluree_connection}.}
}
\value{
A named list of prefixes, or \code{NULL}.
}
\description{
The default JSON-LD context of a connection
}
\examples{
\dontrun{
fluree_context(con)
}
}
