% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{fluree_ledger}
\alias{fluree_ledger}
\title{The ledger a connection points at}
\usage{
fluree_ledger(con)
}
\arguments{
\item{con}{A \code{fluree_connection}.}
}
\value{
The ledger name, as a string.
}
\description{
The ledger a connection points at
}
\examples{
\dontrun{
fluree_ledger(con)
}
}
