% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-builder.R
\name{fluree_query_body}
\alias{fluree_query_body}
\title{The JSON-LD Query a query object will send}
\usage{
fluree_query_body(q)
}
\arguments{
\item{q}{A \code{fluree_query}.}
}
\value{
A list, ready to be serialised as JSON-LD Query.
}
\description{
Assembles the query without sending it. Filters are placed inside \code{where},
because a top-level \code{filter} field is ignored by Fluree v4.1.
}
\examples{
\dontrun{
fluree_query_body(fq_limit(fluree_query(con), 5))
}
}
