% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{fluree_query_json}
\alias{fluree_query_json}
\title{The query as a JSON string}
\usage{
fluree_query_json(q, pretty = TRUE)
}
\arguments{
\item{q}{A \code{fluree_query}.}

\item{pretty}{Whether to indent the output.}
}
\value{
A single string.
}
\description{
The query as a JSON string
}
\examples{
\dontrun{
cat(fluree_query_json(fluree_query(con) |> fq_crawl("?s") |>
                      fq_where(`@id` = "?s", `schema:name` = "?name")))
}
}
