% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions2.R
\name{flureeFetch}
\alias{flureeFetch}
\title{flureeFetch}
\usage{
flureeFetch(path, method, body)
}
\arguments{
\item{path}{Character. The path to the Fluree database.}

\item{body}{List. The body of the request.}

\item{fluree_link}{Character. The link to the Fluree database.}
}
\value{
character
}
\description{
Helper function to fetch data from Fluree
}
