% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeInsert.R
\name{flureeInsert}
\alias{flureeInsert}
\title{flureeInsert}
\usage{
flureeInsert(data, config, signTransaction, apiKey = NULL)
}
\arguments{
\item{data}{list A list representing a JSON-LD object that must be inserted
into a Fluree ledger.}

\item{config}{list The configuration list for the Fluree instance.}

\item{signTransaction}{logical Indicates whether the transaction should be
signed or not.}

\item{apiKey}{character A Fluree API key.}
}
\value{
The output of sendTransaction.
}
\description{
Inserts data into a Fluree ledger.
}
