% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{flureeLedgerRef}
\alias{flureeLedgerRef}
\title{Construct a branch-qualified ledger reference}
\usage{
flureeLedgerRef(ledger, branch = "main")
}
\arguments{
\item{ledger}{Ledger name.}

\item{branch}{Branch name.}
}
\value{
A \code{ledger:branch} reference.
}
\description{
Construct a branch-qualified ledger reference
}
