% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk2pem.R
\name{flureePK2pem}
\alias{flureePK2pem}
\title{flureePK2pem}
\usage{
flureePK2pem(k, file_path = "private_key.pem", writePEM = TRUE)
}
\arguments{
\item{k}{Character. The 64 character private key.}

\item{file_path}{Character. The path to save the PEM file. Default is "private_key.pem".}

\item{writePEM}{Logical. If TRUE, the PEM file will be saved to the file_path.
If FALSE, returns the key as class  "key" "ecdsa". Default is TRUE.}
}

\value{
Writes to file_path and returns a message
}

\description{
Function to convert a Fluree private key to a PEM file.
}
\examples{
flureePK2pem(k, "private_key.pem")
}
