% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeTransact.R
\name{flureeTransact}
\alias{flureeTransact}
\title{flureeTransact}
\usage{
flureeTransact(ledgerName, transactObject, signQuery = TRUE)
}
\arguments{
\item{ledgerName}{Character. The name of the ledger.}

\item{transactObject}{json The object to be transacted.}

\item{signQuery}{Logical. Should the query be signed?}
}
\value{
character
}
\description{
Helper function to transact in Fluree
}
