% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorHandling.R
\name{flureeVector}
\alias{flureeVector}
\title{Construct a native Fluree vector literal}
\usage{
flureeVector(vector, query = FALSE)
}
\arguments{
\item{vector}{Numeric embedding vector.}

\item{query}{Whether the literal will be used in a query. Query literals use
the full \code{f:embeddingVector} datatype because \verb{@vector} is transaction-only
shorthand.}
}
\value{
A JSON-LD typed-value list.
}
\description{
Construct a native Fluree vector literal
}
