% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripples_to_jsonld_helpers.R
\name{format_object}
\alias{format_object}
\title{format_object}
\usage{
format_object(obj, obj_type, context_df = NULL)
}
\arguments{
\item{obj_type}{}
}
\description{
Function to convert object based on type. If the type is "uri", the object should be a list with @id as the name so the the json can be "@id": \if{html}{\out{<object>}}.
}
