% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBodyFunctions.R
\name{generateBodyFunctions}
\alias{generateBodyFunctions}
\title{generateBodyFunctions}
\usage{
generateBodyFunctions(ledgerName = "authority/test", signQuery = FALSE)
}
\arguments{
\item{ledgerName}{Character. The name of the ledger.}

\item{signQuery}{Logical. Should the query be signed?}
}
\value{
character
}
\description{
Helper function to generate the body of a function
}
