% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyHandling.R
\name{generateKeyPair}
\alias{generateKeyPair}
\title{Generate a new key pair.}
\usage{
generateKeyPair()
}
\description{
This method makes use of the flureeCrypto package to generate a new private key.
The public key and DID are then derived from the private key and these are added
to the \code{config} of the current Fluree instance.
}
