% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions2.R
\name{getAllEntityRecords}
\alias{getAllEntityRecords}
\title{getAllEntityRecords}
\usage{
getAllEntityRecords(ledgerName, entityName, signQuery = TRUE)
}
\arguments{
\item{ledgerName}{Character. The name of the ledger.}

\item{entityName}{Character. The name of the entity.}
}
\value{
character
}
\description{
Helper function to fetch all records of an entity from Fluree
}
\examples{
:
dfCollections <- getAllEntityRecords("authority/test", "_collection")
dfUser <- getAllEntityRecords("authority/test", "_user")
dfAuth <- getAllEntityRecords("authority/test", "_auth")
dfPredicate <- getAllEntityRecords("authority/test", "_predicate")
}
