% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextHandling.R
\name{getContext}
\alias{getContext}
\title{Get the context}
\usage{
getContext()
}
\value{
(\code{list()})
}
\description{
This function extracts the \code{defaultContext} from the \code{config} of the current
Fluree instance.
}
