% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getDefaultFromJSONargs}
\alias{getDefaultFromJSONargs}
\title{getDefaultFromJSONargs}
\usage{
getDefaultFromJSONargs()
}
\description{
To be called by any function in this package that makes use of
jsonlite::fromJSON. Ensures fromJSON conversion consistency across all functions.
}
