% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getDefaultToJSONargs}
\alias{getDefaultToJSONargs}
\title{getDefaultToJSONargs}
\usage{
getDefaultToJSONargs(pretty = FALSE)
}
\description{
To be called by any function in this package that makes use of
jsonlite::toJSON. Ensures toJSON conversion consistency across all functions.
}
