% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyHandling.R
\name{getKey}
\alias{getKey}
\title{Retrieve Private Key}
\usage{
getKey()
}
\description{
This method retrieves the stored private key using the \code{key_get()} method of
the \code{keyring} package. Note the private key must have been previously stored
using the \code{setKey()} method.

NOTE: This function should be used with caution. Only invoke this function
withing another function call and never store your private key as an
environment variable.
}
