% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredicateInfo.R
\name{getPredicateInfo}
\alias{getPredicateInfo}
\title{Autofill predicate information}
\usage{
getPredicateInfo(ontology_path, pred_vn, id_tb)
}
\arguments{
\item{ontology_path}{\link{string} Path to local ontology file, preferably in RDF/XML syntax (.rdf extension)}

\item{pred_vn}{\link{list} Named list where the keys are the predicate IRIs and the values are variable names
(NA if no corresponding variable name in your data exists.)}

\item{id_tb}{\link{tibble} Class specifications for your data. You can use \code{\link[=identify_nodes]{identify_nodes()}} to generate this tibble.}
}
\value{
\link{list} Named list with two entries: \code{mapped} and \code{exceptions}
}
\description{
Given a list of predicate IRIs and the variable names they correspond to (possibly NA if there is no corresponding variable),
autofill the domain and range of the predicate. If the domain and/or range class is not in the provided node/class specifications (\code{id_tb}),
it attempts to find a (unique) subclass of the domain and/or range class that is.
}
\details{
This function returns two tibbles: \code{mapped} and \code{exceptions}. \code{mapped} contains the "unproblematic" mappings
and \code{exceptions} those mappings that have problems and need to be specified manually. Manual specification can be done using \code{\link[=mapPredicates]{mapPredicates()}}.

There are four exceptions:
\enumerate{
\item Neither a class nor a unique subclass of that class is present in \code{id_tb}
\item Domain and/or range not specified for an owl:ObjectProperty
\item varname = NA for owl:DatatypeProperty
\item No variable name and no domain provided
}
}
