% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyHandling.R
\name{getPublicKey}
\alias{getPublicKey}
\title{Get the public key}
\usage{
getPublicKey()
}
\value{
(\code{string}) | (\code{undefined}).
}
\description{
Get the currently configured public key (if one has been set).
}
