% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{getQueryText}
\alias{getQueryText}
\title{Get the Raw Query Text}
\usage{
getQueryText(queryVariables = NULL, pretty = TRUE)
}
\value{
JSON string representation of the query body
}
\description{
This function returns the query body as a JSON string.
If the query instance has already been signed,  the signature is deserialized
before returning the raw JSON string.
}
\examples{
\dontrun{
queryInstance <- query(exampleQuery)

qry <- getQueryText(queryInstance)

}
}
