% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{getTransactionText}
\alias{getTransactionText}
\title{Get the transaction}
\usage{
getTransactionText(transactionVariables = NULL)
}
\value{
JSON string representation of the transaction body
}
\description{
This function returns the transaction body as a JSON string.
}
\examples{
\dontrun{
transactionInstance <- transact(exampleData)

txn  <- getTransactionText(transactionInstance)
}

}
