% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{The primary value to check. If it is not null this value is returned.}

\item{b}{The fallback value to return if \code{a} is \code{NULL}.}
}
\value{
The value of \code{a} if it is not \code{NULL}, otherwise the value of \code{b}.
}
\description{
This helper function provides a null-coalescing operator, which returns
the first non-\code{NULL} value between two arguments. It is useful for
providing default values in cases where the first argument might be \code{NULL}.
}
