% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionUtils.R
\name{handleDelete}
\alias{handleDelete}
\title{Handle a delete transaction}
\usage{
handleDelete(id, idAlias)
}
\arguments{
\item{id}{(\code{string})\cr
The entry to be deleted.}

\item{idAlias}{(\code{string})\cr
The alias used for the @id field in the context of the Fluree instance.}
}
\value{
(\code{list})
}
\description{
This function serves to create the transaction body to be sent to the Fluree
instance in order to perform the delete.
}
