% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionUtils.R
\name{handleUpsert}
\alias{handleUpsert}
\title{Handle a delete transaction}
\usage{
handleUpsert(upsertTxn, idAlias)
}
\arguments{
\item{upsertTxn}{(\code{list()})\cr
The upsert transaction to rewrite.}

\item{idAlias}{(\code{string})\cr
The alias used for the @id field in the context of the Fluree instance.}
}
\value{
(\code{list})
}
\description{
This function serves to create the transaction body to be sent to the Fluree
instance in order to perform the delete.
}
