% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoryQueryInstance.R
\docType{class}
\name{HistoryQueryInstance}
\alias{HistoryQueryInstance}
\title{Class for querying the commit log of a Fluree ledger.}
\description{
Class for querying the commit log of a Fluree ledger.

Class for querying the commit log of a Fluree ledger.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{query}}{(\code{list()})\cr
Optional parameters: \code{limit} (integer), \code{from} (start t), \code{to} (end t).}

\item{\code{config}}{(\code{list()})\cr
Configuration parameters of the instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HistoryQueryInstance-new}{\code{HistoryQueryInstance$new()}}
\item \href{#method-HistoryQueryInstance-send}{\code{HistoryQueryInstance$send()}}
\item \href{#method-HistoryQueryInstance-getQuery}{\code{HistoryQueryInstance$getQuery()}}
\item \href{#method-HistoryQueryInstance-clone}{\code{HistoryQueryInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistoryQueryInstance-new"></a>}}
\if{latex}{\out{\hypertarget{method-HistoryQueryInstance-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistoryQueryInstance$new(query = list(), config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{(\code{list()})\cr
Optional parameters for the log request: \code{limit}, \code{from} (start t), \code{to} (end t).}

\item{\code{config}}{(\code{list()})\cr
Configuration parameters of the instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistoryQueryInstance-send"></a>}}
\if{latex}{\out{\hypertarget{method-HistoryQueryInstance-send}{}}}
\subsection{Method \code{send()}}{
Fetch the commit log for the configured ledger.
Returns a list with \code{ledger_id}, \code{commits} (each having \code{t}, \code{commit_id}, \code{time},
\code{asserts}, \code{retracts}, \code{flake_count}), \code{count}, and \code{truncated}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistoryQueryInstance$send()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistoryQueryInstance-getQuery"></a>}}
\if{latex}{\out{\hypertarget{method-HistoryQueryInstance-getQuery}{}}}
\subsection{Method \code{getQuery()}}{
Returns the query parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistoryQueryInstance$getQuery()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{list()}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistoryQueryInstance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HistoryQueryInstance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistoryQueryInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
