% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodify.R
\name{identify_nodes}
\alias{identify_nodes}
\title{Identify RDF nodes from dataframe and provided schema}
\usage{
identify_nodes(node_spec, data)
}
\arguments{
\item{node_spec}{\code{list} Node specification as a list of named lists. There should be an entry for each class of entity implied by the dataframe.
Each named list should have the following fields:

a) \code{type}: \code{rdf:type} for each node according to nova-o
b) \code{id_col}: the variable that can be used as ID for the node.
c) \code{comb_id_col}: if b) is NULL, which combination of columns can be used to create an ID for the node. The specified columns are hashed to create the ID.
d) \code{const_id}: if both b) and c) are null, the constant identifier for this node. For example, the identifier for the single survey of which all responses form part.

Exactly one of b), c) or d) must be non-NULL}

\item{data}{\code{tibble}}
}
\value{
\code{list} \code{data} with ID columns appended; a tibble containing the node specifications from \code{node_spec} + the variable name used to identify each node (i.e. the subject in each triple)
}
\description{
This function creates node specifications implied by a dataframe according to a specified schema and the Nova ontology.
It amends \code{data} to include columns that are used as the the subject for triples concerning each class.
}
\examples{
\dontrun{
interview_spec <- list(
  type = "https://nova.org.za/nova-o#Interview",
  id_col = "instanceid",
  comb_id_col = NULL,
  const_id = NULL
)

hh_spec <- list(
  type = "https://nova.org.za/nova-o#Household",
  id_col = NULL,
  comb_id_col = c("village", "stand_number_1", "respondent_surname"),
  const_id = NULL
)

survey_spec <- list(
  type = "https://nova.org.za/nova-o#Survey",
  id_col = NULL,
  comb_id_col = NULL,
  const_id = "KiA_adaptation_Q"
)

node_spec <- list(interview_spec, hh_spec, hh_addr_spec, survey_spec)
result <- identify_nodes(node_spec, small_kia_data)
small_kia_data <- result$data
id_tb <- result$id_tb
}
}
