% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeTransact.R
\name{insertData}
\alias{insertData}
\title{insertData}
\usage{
insertData(path, data_list, fluree_link = Sys.getenv("fluree_link"))
}
\arguments{
\item{path}{Character. The path to the Fluree database.}

\item{data_list}{List. The schema to be created (predicates).}

\item{fluree_link}{Character. The link to the Fluree database.}
}
\value{
character
}
\description{
Helper function to create a schema in Fluree
}
