% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{is_fluree_connection}
\alias{is_fluree_connection}
\title{Is this a Fluree connection?}
\usage{
is_fluree_connection(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is this a Fluree connection?
}
\examples{
is_fluree_connection(1)
}
