% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-builder.R
\name{is_fluree_query}
\alias{is_fluree_query}
\title{Is this a Fluree query?}
\usage{
is_fluree_query(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is this a Fluree query?
}
\examples{
is_fluree_query(1)
}
