% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_from_tibble.R
\name{jsonld_from_rdf}
\alias{jsonld_from_rdf}
\title{jsonld_from_rdf}
\usage{
jsonld_from_rdf(rdf)
}
\arguments{
\item{rdf}{rdf object}

\item{format}{Character. Default "application/ld+json"}
}
\value{
jsonld object
}
\description{
jsonld_from_rdf
}
\examples{
\dontrun{
jsonld_from_rdf(rdf1)
}
}
