% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchHandling.R
\name{listBranches}
\alias{listBranches}
\title{List branches of a Fluree ledger}
\usage{
listBranches(config, ledger = config$ledger)
}
\arguments{
\item{config}{Fluree configuration created by \code{\link[=setConfig]{setConfig()}}.}

\item{ledger}{Ledger name. Defaults to \code{config$ledger}.}
}
\value{
A list of branch records returned by Fluree.
}
\description{
List branches of a Fluree ledger
}
