% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{make_extended_cto_semantic_mapping}
\alias{make_extended_cto_semantic_mapping}
\title{make_extended_cto_semantic_mapping}
\usage{
make_extended_cto_semantic_mapping()
}
\description{
The \code{make_extended_cto_semantic_mapping} function creates an extended mapping between SurveyCTO field types and their corresponding semantic labels.
}
