% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{make_surveycto_centext}
\alias{make_surveycto_centext}
\title{make_surveycto_centext}
\usage{
make_surveycto_centext()
}
\description{
This function creates a context for SurveyCTO RDF output. It defines the prefixes and namespaces used in the RDF triples.
}
