% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_surveycto_context.R
\name{make_surveycto_context}
\alias{make_surveycto_context}
\title{make_surveycto_context}
\usage{
make_surveycto_context()
}
\value{
tibble with columns \code{prefix} and \code{namespace}
}
\description{
make_surveycto_context
}
\examples{
surveycto_context <- make_surveycto_context()
}
