% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_surveycto_context.R
\name{make_surveycto_context_list}
\alias{make_surveycto_context_list}
\title{make_surveycto_context_list}
\usage{
make_surveycto_context_list()
}
\value{
named list with prefixes as names and namespaces as values
}
\description{
Create a named list from the surveycto context tibble
}
