% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQuerySignature.R
\name{makeQuerySignature}
\alias{makeQuerySignature}
\title{makeQuerySignature}
\usage{
makeQuerySignature(
  ledgerName,
  privateKey = Sys.getenv("privateKey"),
  queryString,
  authId = Sys.getenv("authId")
)
}
\arguments{
\item{ledgerName}{Character. The name of the ledger.}

\item{privateKey}{Character. The private key.}

\item{queryString}{Character. The query string.}

\item{authId}{Character. The authId.}
}
\value{
character
}
\description{
Helper function to sign a query in Fluree
}
