% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post2fluree.R
\name{makeTransaction}
\alias{makeTransaction}
\title{makeTransaction}
\usage{
makeTransaction(id = NULL, transaction_type = c("insert", "delete"), data)
}
\arguments{
\item{id}{Character. The id of the object to be inserted or deleted.}

\item{transaction_type}{Character. The type of transaction. Default is "insert".}

\item{data}{Data frame with two columns: Key and Value. The data to be inserted or deleted.
The first column is the key and the second column is the value.}
}
\description{
Make a transaction_data object to be sent to Fluree.
}
