% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPredicates.R
\name{mapPredicates}
\alias{mapPredicates}
\title{Create specified predicate mappings from dataframe}
\usage{
mapPredicates(varnames, predIRIs, domains, ranges)
}
\arguments{
\item{varnames}{\link{character} Column names corresponding to predicates.}

\item{domains}{\link{character} Complete domain IRIs.}

\item{ranges}{\link{character} Complete range IRIs.}

\item{predicates}{\link{character} Complete predicate IRIs.}
}
\value{
\link{tibble} The predicate mappings as a tibble with columns:
\code{predicate}, \verb{http://www.w3.org/2000/01/rdf-schema#domain},
\verb{http://www.w3.org/2000/01/rdf-schema#range}, and \code{varname} (column name corresponding to this predicate mapping).
}
\description{
This function maps the columns of a dataframe to predicate specifications using conventions compatible with Fluree V3.
These predicates can be used to define a schema for Fluree transactions.
}
\details{
Each predicate mapping specifies:
\enumerate{
\item \strong{predicate}: the IRI of the predicate in full. Inclusion of \verb{@context} to incorporate ontology prefixes is handled later in the data pipeline.
\item \strong{domain}: the domain of the predicate as an IRI.
\item \strong{range}: the range of the predicate as an IRI.
\item \strong{varname}: the corresponding column (variable) name in the dataframe.
}

This function just builds the predicate specification.
The heavy lifting (the actual content of the specification) still has to be done manually.
}
