% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCTOMetadata.R
\name{parseCTO}
\alias{parseCTO}
\title{Autofill standard triples for SurveyCTO triples}
\usage{
parseCTO(
  data,
  instanceid = "instanceid",
  starttime = "starttime",
  endtime = "endtime",
  devicephonenum = "devicephonenum",
  device_info = "device_info",
  duration = "duration",
  geo_location = "gps_location",
  village = "village",
  standnumber = "stand_number_1",
  survey,
  ...
)
}
\arguments{
\item{data}{\link{data.frame} The survey response data}

\item{instanceid}{\link{string} Column in \code{data} that contains ID for interview}

\item{starttime}{\link{string} Column in \code{data} that contains starting time for interview}

\item{endtime}{\link{string} Column in \code{data} that contains ending time for interview}

\item{devicephonenum}{\link{string} Column in \code{data} that contains phone number for the device used to conduct the survey}

\item{device_info}{\link{string} Column in \code{data} that contains information on the device used to conduct the interview}

\item{duration}{\link{string} Column in \code{data} that contains duration of the interview}

\item{geo_location}{\link{string} Column in \code{data} that contains location at which interview was conducted}

\item{village}{\link{string} Column in \code{data} that contains the village respondent lives in}

\item{standnumber}{\link{string} Column in \code{data} that contains the respondent's stand number}

\item{survey}{\link{string} Identifier for the survey these responses form part of.}
}
\value{
\link{data.frame} Containing columns \code{subject}, \code{predicate}, \code{object}. This dataframe can be parsed to triples.
}
\description{
TODO not complete
}
\details{
This function:

a) Constructs node (class) specifications compliant with \code{\link[=identify_nodes]{identify_nodes()}} for the following classes from nova-o:
\code{Interview}, \code{InterviewResponse}, \code{Survey}, \code{InterviewDevice}, \code{HouseholdAddress}, and \code{Household}

b) Constructs, where necessary, additional \verb{<classname>_ID} variables used to identify nodes (instances of classes) above

c) Constructs predicate mappings for the following predicates:
\verb{http://www.w3.org/ns/prov#startedAtTime},
\verb{http://www.w3.org/ns/prov#endedAtTime},
\verb{https://www.w3.org/2006/vcard/ns#hasTelephone},
\verb{https://nova.org.za/nova-o#hasDeviceInfo},
\verb{http://purl.org/aiaontology#hasDuration},
\verb{http://purl.org/aiaontology#hasLocation},
\verb{https://nova.org.za/nova-o#inVillage},
\verb{https://nova.org.za/nova-o#hasStandNumber},
\verb{https://nova.org.za/nova-o#conductedWithDevice},
\verb{https://nova.org.za/nova-o#isResponseOf},
\verb{http://purl.org/aiaontology#usedToPerform},
\verb{https://nova.org.za/nova-o#interviewWithHousehold},
\verb{https://nova.org.za/nova-o#hasAddress}

d) Maps \code{data} to SPO format using the above class specifications and predicates.
}
