% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer_with_type.R
\name{pivot_longer_with_type}
\alias{pivot_longer_with_type}
\title{pivot_longer_with_type}
\usage{
pivot_longer_with_type(data, names_to = "name", tripplenames = TRUE, ...)
}
\arguments{
\item{names_to}{Character. What you want to call the attribute column. Default "name"}

\item{tripplenames}{Logical. Default TRUE. If TRUE, name columns subject, predicate, object and type}
}
\description{
pivot_longer_with_type
}
