% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer_with_type.R
\name{pivot_wider_by_type}
\alias{pivot_wider_by_type}
\title{pivot_wider_by_type}
\usage{
pivot_wider_by_type(
  res,
  subject = "subject",
  predicate = "predicate",
  object = "object"
)
}
\arguments{
\item{res}{data.frame}

\item{subject}{Character. Name of column in res with subject. Default "subject"}

\item{predicate}{Character. Name of column in res with predicate. Default "predicate"}

\item{object}{Character. Name of column in res with object Default "object"}
}
\description{
pivot_wider_by_type
}
\examples{
df1 = pivot_longer_with_type(iris)
df1 \%>\% dplyr::select(predicate, type) \%>\% dplyr::distinct()
df2 = pivot_wider_by_type(df1)
}
