% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSPO.R
\name{pivotLongerSPO}
\alias{pivotLongerSPO}
\title{Create triples from dataframe}
\usage{
pivotLongerSPO(data, pred_tb)
}
\arguments{
\item{data}{\link{data.frame}}

\item{pred_tb}{\code{tibble} A tibble containing columns \code{predicate} (which contains the prefixed IRI), \code{rdfs:domain}, \code{rdfs:range} and \code{varname} (which contains the name of the variable in data corresponding to this predicate)
You can use \code{\link[=mapPredicates]{mapPredicates()}} and \code{\link[=predicateTibble]{predicateTibble()}} to generate this tibble.}
}
\value{
A dataframe containing columns \code{subject}, \code{predicate}, \code{object}
}
\description{
This function parses the data from the parameter \code{data} and ensures compliance with the predicate mappings in \code{pred_tb},
i.e. it ensures that the subject and object for each triple is correct.
}
\details{
It expects a dataframe containing additional columns \verb{<class_name>_ID} for each class implied by the dataframe,
most likely generated by \code{\link[=identify_nodes]{identify_nodes()}}. These ID columns contain the values that will be used as subjects in resulting triples.
}
\examples{
\dontrun{
pred_tb <- tibble(
  predicate = c("prov:startedAtTime", "prov:endedAtTime", "rdf:type"),
  `rdfs:domain` = c(":Interview", ":Interview", "owl:Thing"),
  `rdfs:range` = c("xsd:dateTime", "xsd:dateTime", ":Interview"),
  varname = c("starttime", "endtime", "Interview_ID")
)

small_data <- tibble(
  starttime = c("2024-02-27 14:18:19", "2024-02-27 09:40:44"),
  endtime = c("2024-02-27 14:41:43", "2024-02-27 09:59:54"),
  Interview_ID = c("uuid:fe690b51-ee58-4294-b7b6-9f941b47b741",
                   "uuid:0aa90b5e-14a7-46bc-a4cd-dc7306f03d61")
)

pivotLongerSPO(small_data, pred_tb)
}
}
\seealso{
\code{\link[=mapPredicates]{mapPredicates()}}, \code{\link[=predicateTibble]{predicateTibble()}}
}
