% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rdf_triples_interactive.R
\name{plot_rdf_triples_interactive}
\alias{plot_rdf_triples_interactive}
\title{plot_rdf_triples_interactive}
\usage{
plot_rdf_triples_interactive(triples_df, context_map)
}
\arguments{
\item{context_map}{}
}
\description{
Take the dataframe of tripples made by map_cto_to_rdf and create a graph.
}
\examples{
\dontrun{
kia_adapt <- novaCTO::readCTO("KiA_adaptation_ACTIVE")
surveycto_context <- make_surveycto_context()
formdef <- kia_adapt$fromschema$kia_adaptation
fromRDF <- map_cto_to_rdf(formdef, base_uri = "https://novapc.surveycto.com/", instrument = "KiA_adaptation_ACTIVE")
p1 <- plot_rdf_triples_interactive(fromRDF, surveycto_context)
p1
htmlwidgets::saveWidget(p1, file = "~/tmp/rdf_graph.html", selfcontained = TRUE)
utils::browseURL("~/tmp/rdf_graph.html")
}
}
