% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post2fluree.R
\name{post2fluree}
\alias{post2fluree}
\title{post2fluree}
\usage{
post2fluree(
  signature = NULL,
  privateKey = NULL,
  tx_type = "transact",
  ledger,
  transaction_data,
  fluree_link = "http://localhost:8095/",
  endpoint = "fluree/create"
)
}
\arguments{
\item{signature}{Character. The signature to be included in the headers. The the result
of the sign_query function.}

\item{privateKey}{Character. The private key to be used to sign the query. Either a .pem file or of class "key" or "ecdsa".}

\item{tx_type}{Character. The type of transaction. Default is "transact" # check fluree docs for types}

\item{ledger}{Character. The name of the ledger.}

\item{transaction_data}{Character. The data to be sent to Fluree. JSON string.}

\item{fluree_link}{Character. The URL of the Fluree server.}

\item{endpoint}{Character. The endpoint to be used. Default is "fluree/create".}
}
\value{
Response from the POST request.
}
\description{
POST transaction data to Fluree.
}
\examples{
transaction_data <- list(`@context` = list(ex = "http://example.org/", schema = "http://schema.org/"),
ledger = "cookbook/base",
insert = structure(list(`@id` = "ex:freddy", `@type` = "ex:Yeti", `schema:age` = 4L, `schema:name` = "Freddy",  `ex:verified` = TRUE),
class = "data.frame", row.names = 1L))
pk <- "private_key.pem"
p1 <- post2fluree(signature = NULL, privateKey = pk, tx_type = "transact", ledger = "cookbook/base", transaction_data = transaction_data)
}
