% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates_from_df.R
\name{predicates_from_df}
\alias{predicates_from_df}
\title{predicates_from_df}
\usage{
predicates_from_df(data, returnJSON = FALSE)
}
\arguments{
\item{data}{Data frame or tibble}
}
\value{
list
}
\description{
predicates_from_df
}
\examples{
predicates_from_df(iris, returnJSON = T)
predicates_from_df(iris)
}
