% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPredicates.R
\name{predicateTibble}
\alias{predicateTibble}
\title{Convert predicate mapping list to a tibble}
\usage{
predicateTibble(schema_list)
}
\arguments{
\item{schema_list}{Predicate mapping list of lists, most likely generated by \code{mapPredicates}}
}
\value{
Predicate mapping as a tibble. Columns: predicate IRI, domain, range, corresponding column name
}
\description{
Convert predicate mapping list to a tibble
}
\examples{
small_predlist <- mapPredicates(varnames, predIRIs, domains, ranges)
pred_tb <- predicateTibble(small_predlist)
}
