% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(dfData = NULL, collectionName = NULL)
}
\arguments{
\item{dfData}{Dataframe or tibble with the data to be added.}

\item{collectionName}{Character. The name of the collection.}
}
\value{
character
}
\description{
Helper function to add data in Fluree
}
