% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSchema.R
\name{prepareSchema}
\alias{prepareSchema}
\title{prepareSchema}
\usage{
prepareSchema(predicateNames = NULL, dfData = NULL)
}
\arguments{
\item{predicateNames}{List. The names of the predicates.}

\item{dfData}{Dataframe. The data to be added.}
}
\value{
character
}
\description{
Helper function to prepare a schema in Fluree
}
