% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionUtils.R
\name{processEntity}
\alias{processEntity}
\title{Process an entity and build a structured map}
\usage{
processEntity(entity, map, idAlias)
}
\arguments{
\item{entity}{(\code{list})\cr
The entity to process.}

\item{map}{(\code{list})\cr
A map to store processed entities.}

\item{idAlias}{(\code{character})\cr
The alias used for the @id field in the context of the Fluree instance.}
}
\value{
(\code{list})
}
\description{
Recursively processes an entity, extracting nested entities and structuring them
into a map where each entity is keyed by its ID.
}
