% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties2kv.R
\name{properties2kv}
\alias{properties2kv}
\title{properties2kv}
\usage{
properties2kv(dfprops, id, typename = "object_type", to_type = NULL)
}
\arguments{
\item{dfprops}{A data frame with columns \code{predicate}, \code{object}, and \code{object_type} (either "uri" or "literal").}
}
\value{
A named list representing the key-value map.
}
\description{
Convert a properties tibble to a key-value map suitable for JSON-LD serialization.
}
\examples{
\dontrun{
dfprops <- structure(list(predicate = c("rdf:type", "rdfs:label", "survey:hasQuestion", "survey:hasQuestion", "survey:hasQuestion"), object = c("survey:Survey", "SurveyCTO Form", "https://novapc.surveycto.com/KiA_adaptation_ACTIVE/question/starttime", "https://novapc.surveycto.com/KiA_adaptation_ACTIVE/question/endtime", "https://novapc.surveycto.com/KiA_adaptation_ACTIVE/question/deviceid"), object_type = c("uri", "literal", "uri", "uri", "uri")), row.names = c(NA, -5L), class = c("tbl_df", "tbl", "data.frame"))
res <- properties2kv(dfprops, id = "https://novapc.surveycto.com/KiA_adaptation_ACTIVE_partial")
cat(toJSON(res, auto_unbox = TRUE, pretty = TRUE))
}
}
