% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{Query}
\alias{Query}
\title{Configure and Send a Query}
\usage{
Query(...)
}
\arguments{
\item{...}{Passed to \code{\link[=query]{query()}}.}
}
\description{
This convenience function configures and sends the passed query by calling
the relevant functions.
}
\examples{
\dontrun{
# Existing data:
#  [
#    { "@id": "freddy", "name": "Freddy" },
#    { "@id": "alice", "name": "Alice" }
#  ]

exampleQuery <- '{
     "select": {
       "?s": ["*"]
     },
     "where": {
       "@id": "?s",
       "name": "?name"
     }
}'

Query(config = conf, ledger = 'demo', exampleQuery, signQuery = FALSE)

# OR ALTERNATIVELY

queryList <- fromJSON(exampleQuery, simplifyDataFrame = FALSE, simplifyMatrix = FALSE, simplifyVector = FALSE)
Query(config = conf, ledger = 'demo', queryList, signQuery = FALSE)
}

}
\seealso{
\code{\link[=query]{query()}}

\code{\link[=sendQuery]{sendQuery()}}
}
