% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{query}
\alias{query}
\title{Query Configuration}
\usage{
query(
  config = NULL,
  ledger = NULL,
  query,
  signQuery = NULL,
  privateKey = NULL,
  apiKey = NULL
)
}
\arguments{
\item{config}{(\code{list()})\cr
The configuration list for the Fluree instance.}

\item{ledger}{(\code{character})\cr
The name of the ledger to query.}

\item{query}{(\code{list()})\cr
The list representation of the query body to be sent.
Alternatively the query can simply be passed as a JSON \code{character} string.}

\item{signQuery}{(\code{logical})\cr
Determines whether the given query should be signed or not.
This will override the value set in \code{config$signMessages}.}

\item{privateKey}{(\code{character})\cr
The hexstring representation of the private key to use for message signing.
Overrides \code{getKey()} if provided.}

\item{apiKey}{(\code{character})\cr
The API key for accessing the Fluree ledger on the remote host (e.g., when
the host is data.flur.ee).}
}
\value{
A list containing everything needed to query the Fluree database.
This includes all the necessary parameters as well as the signed/unsigned query itself.
}
\description{
This function configures the query instance.
The default context is added and if applicable the query is signed.
The function returns a list referred to as \code{queryVariables} which is a single
structure containing all the information necessary to interact with the
Fluree instance via \code{sendQuery()}.
}
\examples{
\dontrun{
# Existing data:
#  [
#    { "@id": "freddy", "name": "Freddy" },
#    { "@id": "alice", "name": "Alice" }
#  ]

exampleQuery <- '{
     "select": {
       "?s": ["*"]
     },
     "where": {
       "@id": "?s",
       "name": "?name"
     }
}'

queryInstance <- query(exampleQuery)

# OR ALTERNATIVELY

queryList <- fromJSON(exampleQuery, simplifyDataFrame = FALSE, simplifyMatrix = FALSE, simplifyVector = FALSE)
queryInstance <- query(queryList)

}
}
