% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryInstance.R
\docType{class}
\name{QueryInstance}
\alias{QueryInstance}
\title{Class providing objects with methods to query a Fluree instance.}
\description{
Class providing objects with methods to query a Fluree instance.

Class providing objects with methods to query a Fluree instance.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{query}}{(\code{list()})\cr
The list representation of a query to be sent to the Fluree instance.}

\item{\code{config}}{(\code{list()})\cr
Configuration parameters of the instance.}

\item{\code{signedQuery}}{(\code{string})\cr
The JWT of the query.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QueryInstance-new}{\code{QueryInstance$new()}}
\item \href{#method-QueryInstance-send}{\code{QueryInstance$send()}}
\item \href{#method-QueryInstance-sign}{\code{QueryInstance$sign()}}
\item \href{#method-QueryInstance-getSignedQuery}{\code{QueryInstance$getSignedQuery()}}
\item \href{#method-QueryInstance-getQuery}{\code{QueryInstance$getQuery()}}
\item \href{#method-QueryInstance-clone}{\code{QueryInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryInstance-new"></a>}}
\if{latex}{\out{\hypertarget{method-QueryInstance-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryInstance$new(query, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{(\code{list()})\cr
The query to be sent to the Fluree instance.}

\item{\code{config}}{(\code{list()})\cr
Configuration parameters of the instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryInstance-send"></a>}}
\if{latex}{\out{\hypertarget{method-QueryInstance-send}{}}}
\subsection{Method \code{send()}}{
This method sends the configured query to the host.
The Fluree instance must be 'connected' before querying or transacting.
If \code{signMessages = TRUE} the JWT will be sent to the host.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryInstance$send()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryInstance-sign"></a>}}
\if{latex}{\out{\hypertarget{method-QueryInstance-sign}{}}}
\subsection{Method \code{sign()}}{
This method is used to sign the configured query.
This method is called automatically when \code{signMessages = TRUE} and requires
a privateKey to be either passed as a parameter or configured within the
\code{config}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryInstance$sign(privateKey = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{privateKey}}{(\code{string})\cr
The private key to use for message signing (represented as a hex string).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{QueryInstance}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryInstance-getSignedQuery"></a>}}
\if{latex}{\out{\hypertarget{method-QueryInstance-getSignedQuery}{}}}
\subsection{Method \code{getSignedQuery()}}{
Returns the signed query (if it has been set).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryInstance$getSignedQuery()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{string}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryInstance-getQuery"></a>}}
\if{latex}{\out{\hypertarget{method-QueryInstance-getQuery}{}}}
\subsection{Method \code{getQuery()}}{
Returns the query body.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryInstance$getQuery()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{string}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryInstance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QueryInstance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
