% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_from_tibble.R
\name{rdf_from_df3}
\alias{rdf_from_df3}
\title{rdf_from_df3}
\usage{
rdf_from_df3(
  df,
  subject = "subject",
  predicate = "predicate",
  object = "object",
  base = NULL,
  vocab = NULL
)
}
\arguments{
\item{df}{Dataframe or tibble}

\item{predicate}{Character. Default "predicate"}

\item{object}{Character. Default "object"}

\item{subject.}{Character. Default "subject"}
}
\value{
rdf object
}
\description{
rdf_from_df3
}
\examples{
\dontrun{
iristriples <- pivot_longer_with_type(iris)
rdfiris <- rdf_from_df3(iristriples, base = "http://example.com/iris/", vocab = "http://example.com/irisvocab/#")
rdf_serialize(rdfiris, "rdfiris.json", format = "jsonld")
}
}
