% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextHandling.R
\name{removeDuplicateUnnamed}
\alias{removeDuplicateUnnamed}
\title{Remove Duplicate Unnamed Context Elements}
\usage{
removeDuplicateUnnamed(contextList)
}
\arguments{
\item{contextList}{The context list with potential duplicates.}
}
\value{
The context list with duplicates removed.
}
\description{
Removes duplicate unnamed elements from a context list.
}
