% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacePrefixes.R
\name{replace_iris_with_prefixes}
\alias{replace_iris_with_prefixes}
\title{Replace IRIs with prefixes}
\usage{
replace_iris_with_prefixes(data, add_context = NULL)
}
\arguments{
\item{data}{Predicate mappings. The function expects a list of lists, most likely generated by \code{mapPredicates}}

\item{add_context}{Additional prefix mappings as a named list.}
}
\value{
Prefixed predicate mappings
}
\description{
This function courtesy of Claude.
This function replaces verbose IRIs in predicate mappings with their prefixed versions.
To do this, it uses the context list below, as well as additional prefix mappings provided in @param add_context
context <- list(
":" = "https://nova.org.za/nova-o#",
"aia" = "http://purl.org/aiaontology#",
"owl" = "http://www.w3.org/2002/07/owl#",
"rdf" = "http://www.w3.org/1999/02/22-rdf-syntax-ns#",
"xml" = "http://www.w3.org/XML/1998/namespace",
"xsd" = "http://www.w3.org/2001/XMLSchema#",
"foaf" = "http://xmlns.com/foaf/spec/#term_",
"prov" = "http://www.w3.org/ns/prov#",
"rdfs" = "http://www.w3.org/2000/01/rdf-schema#",
"time" = "https://www.w3.org/2006/time#",
"vcard" = "http://www.w3.org/TR/vcard-rdf/#",
"base" = "https://nova.org.za/nova-o"
)
}
\examples{
small_predlist <- mapPredicates(varnames, predIRIs, domains, ranges)
geprefix <- replace_iris_with_prefixes(small_predlist)
}
