% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredicateInfo.R
\name{replaceSubclass}
\alias{replaceSubclass}
\title{Replace domain/range with unique subclass}
\usage{
replaceSubclass(rdf_data, ex_classes, id_tb)
}
\arguments{
\item{rdf_data}{\code{rdf} Parsed ontology data}

\item{ex_classes}{\code{tibble} Generated with \code{\link[=excludedClasses]{excludedClasses()}}}

\item{id_tb}{\code{tibble} Class specifications for your data. You can use \code{\link[=identify_nodes]{identify_nodes()}} to generate this tibble.}
}
\value{
\code{tibble} Containing columns \code{new_domain} and \code{new_range} (possibly NA)
}
\description{
Helper function for \code{\link[=getPredicateInfo]{getPredicateInfo()}}
For each class in the domain and/or range of a list of predicates
that are not in the node specifications (\code{id_tb}) for a dataset (\code{ex_classes}),
attempt to find a subclass of that class that is in \code{id_tb} such that this subclass is unique (no conflicts).
}
