% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{runDockerContainer}
\alias{runDockerContainer}
\title{Function to run the docker container}
\usage{
runDockerContainer(
  port = "58090",
  name = NULL,
  dockerImage = "5839ffe273062b8da972b120deb54dd62e7c3d1f"
)
}
\arguments{
\item{port}{(\code{string})\cr
The port where the instance is running (e.g. '58090')}

\item{name}{(\code{string})\cr
The name to be linked to the docker container.}

\item{dockerImage}{(\code{string})\cr
The current stable version of the fluree/server docker image.}
}
\description{
This function takes flags as parameter inputs and adds these flags to the
container when the docker image is run.
}
