% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{sendHistoryQuery}
\alias{sendHistoryQuery}
\title{Send a History Query}
\usage{
sendHistoryQuery(queryVariables)
}
\arguments{
\item{queryVariables}{(\code{list()})\cr
A list representing the history query specifications.}
}
\value{
A character string containing the response content.
}
\description{
This function makes use of the \code{httr} package to send the configured history
query to the Fluree \verb{/history} API endpoint.
}
\examples{
\dontrun{
historyQueryInstance <- history(exampleHistoryQuery)
sendHistoryQuery(historyQueryInstance)
}

}
