% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setContext.R
\name{set_context}
\alias{set_context}
\title{Set the context for the survey ontology.}
\usage{
set_context()
}
\value{
JSON object containing the base context of the ontology.
}
\description{
This method sets the base context used in the Survey Ontology.  Which can be
found here: https://cefriel.github.io/survey-ontology/docs/index.html
}
