% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startUp.R
\name{setConfiguration}
\alias{setConfiguration}
\title{Set the configuration for the Fluree instance}
\usage{
setConfiguration(config = list())
}
\arguments{
\item{config}{(\code{list()})\cr
The parameters to configure the Fluree instance with.}
}
\description{
This function validates the configuration parameters.
Once checked the configuration parameters are added to the system environment.
}
\examples{
config <- list(host = "localhost", port = "58090", ledger = "test", create = TRUE)

}
