% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextHandling.R
\name{setContext}
\alias{setContext}
\title{Set the default context}
\usage{
setContext(context)
}
\description{
The default context set here will be used for all queries and transactions.
Unlike \code{addToContext()} this method does not merge new context elements
with existing ones, instead it will replace the existing
\code{defaultContext} entirely.
}
