% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyHandling.R
\name{setKey}
\alias{setKey}
\title{Store Private Key}
\usage{
setKey()
}
\description{
This function securely stores a private key in the system keyring
using the \code{key_set()} function from the \code{keyring} package.
The key is stored under service name \code{"privateKey"} and keyring \code{"Fluree"}.

NOTE: This function will prompt for input.
}
