% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyHandling.R
\name{setKeys}
\alias{setKeys}
\title{Set keys for encryption}
\usage{
setKeys(privateKey)
}
\arguments{
\item{privateKey}{(\code{string})\cr
The private key to use for message signing as a hex string.}
}
\description{
Add a private key to use for signing.
This key will be added to the config of the current instance and  will also
be used to sign messages by default when using the \code{signTransaction()} or
\code{signQuery()} functions on any future queries or transactions.
The public key and DID will be derived from this private key and added to
the \code{config} of the current Fluree instance as well.
}
