% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripples_to_jsonld_helpers.R
\name{shorten_predicate}
\alias{shorten_predicate}
\title{shorten_predicate}
\usage{
shorten_predicate(pred, context_df = make_surveycto_centext())
}
\arguments{
\item{pred}{Character string of the predicate URI}
}
\description{
Function to convert predicate to short form
}
