% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_query.R
\name{sign_query}
\alias{sign_query}
\title{sign_query}
\usage{
sign_query(private_key, transaction_data, tx_type = "transact", ledger)
}
\arguments{
\item{private_key}{Character. The private key as class ""key" "ecdsa" or the path to a .pem file.
You can transform the private key to a .pem file using the function pk2pem.}

\item{transaction_data}{Character. The data to be signed. JSON string.}

\item{tx_type}{Character. The type of transaction. Default is "transact" # check fluree docs for types}

\item{ledger}{Character. The name of the ledger.}
}
\value{
character
}
\description{
Function to sign the query
}
\references{
https://stackoverflow.com/questions/66554462/turn-rsa-public-and-private-keys-into-strings-in-r
}
