% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureText.R
\name{signatureText}
\alias{signatureText}
\title{description: This function generates the signature text for the query}
\usage{
signatureText(ledgerName, privateKey, endpoint, body, authId)
}
\arguments{
\item{ledgerName}{Character. The name of the ledger.}

\item{privateKey}{Character. The private key.}

\item{body}{Character. The body of the query.}

\item{authId}{Character. The authId.}
}
\value{
character
}
\description{
description: This function generates the signature text for the query
}
