% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPredicates.R
\name{specIDPredicates}
\alias{specIDPredicates}
\title{Create predicate mappings for rdf:type predicates}
\usage{
specIDPredicates(id_tb)
}
\arguments{
\item{id_tb}{\link{data.frame} Containing column \code{var_id_name} with the names of the columns that contain the identifiers for each class.
You can generate \code{id_tb} from node specifications by using \code{identify_nodes}.}
}
\value{
\link{tibble} Containing columns:
\code{predicate} (which will be \code{rdf:type}), \verb{http://www.w3.org/2000/01/rdf-schema#domain},
\verb{http://www.w3.org/2000/01/rdf-schema#range}, and \code{varname} (column name corresponding to this predicate mapping).
}
\description{
Add a \code{rdf:type} predicate for each class implied by \code{id_tb}.
}
