% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{stopDockerContainer}
\alias{stopDockerContainer}
\title{Function to stop the docker container}
\usage{
stopDockerContainer(name = NULL)
}
\arguments{
\item{name}{(\code{string})\cr
The name of the container to be stopped.
(Note this could also be the container ID).}
}
\description{
This function stops the specified docker container.
If no container name (or ID) is provided,  the last container that was started
will be stopped.
}
