% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system2tibble.R
\name{system2tibble}
\alias{system2tibble}
\title{system2tibble}
\usage{
system2tibble(chr = system("docker ps -a", intern = TRUE))
}
\arguments{
\item{chr}{Character. The output of a call to system}
}
\value{
tibble
}
\description{
Turn the results of a call to system into a tibble
}
\examples{
system2tibble(system("docker ps", intern = TRUE))
}
