% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{Transact}
\alias{Transact}
\title{Configure and Send a Transaction}
\usage{
Transact(...)
}
\arguments{
\item{...}{Passed to \code{\link[=transact]{transact()}}.}
}
\description{
This convenience function configures and sends the passed transaction
by calling the relevant functions.
}
\examples{
\dontrun{
exampleData <- '{
   "insert": [
      {
        "@id": "freddy",
        "name": "Freddy"
      },
      {
        "@id": "alice",
        "name": "Alice"
      }
   ]
}'

Transact(config = conf, ledger = 'demo', exampleData, signTransaction = FALSE)

# OR ALTERNATIVELY

dataList <- fromJSON(exampleData, simplifyDataFrame = FALSE, simplifyMatrix = FALSE, simplifyVector = FALSE)
Transact(config = conf, ledger = 'demo', dataList, signTransaction = FALSE)
}

}
\seealso{
\code{\link[=transact]{transact()}}

\code{\link[=sendTransaction]{sendTransaction()}}
}
