% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post2fluree.R
\docType{data}
\name{transaction_data}
\alias{transaction_data}
\title{makeTransaction}
\format{
An object of class \code{list} of length 3.
}
\usage{
transaction_data
}
\arguments{
\item{id}{Character. The id of the object to be inserted or deleted.}

\item{transaction_type}{Character. The type of transaction. Default is "insert".}

\item{data}{Data frame with two columns: Key and Value. The data to be inserted or deleted.
The first column is the key and the second column is the value.}
}
\description{
Make a transaction_data object to be sent to Fluree.
}
\keyword{datasets}
