% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransactionInstance.R
\docType{class}
\name{TransactionInstance}
\alias{TransactionInstance}
\title{Class providing objects with methods to transact to a Fluree instance.}
\description{
Class providing objects with methods to transact to a Fluree instance.

Class providing objects with methods to transact to a Fluree instance.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{transaction}}{(\code{list()})\cr
The list representation of a transaction to be sent to the Fluree instance.}

\item{\code{config}}{(\code{list()})\cr
Configuration parameters of the instance.}

\item{\code{signedTransaction}}{(\code{string})\cr
The JWT of the transaction.}

\item{\code{endpoint}}{(\code{string})\cr
The Fluree v4 endpoint: 'insert', 'upsert', or 'update'.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransactionInstance-new}{\code{TransactionInstance$new()}}
\item \href{#method-TransactionInstance-send}{\code{TransactionInstance$send()}}
\item \href{#method-TransactionInstance-sign}{\code{TransactionInstance$sign()}}
\item \href{#method-TransactionInstance-getSignedTransaction}{\code{TransactionInstance$getSignedTransaction()}}
\item \href{#method-TransactionInstance-getTransaction}{\code{TransactionInstance$getTransaction()}}
\item \href{#method-TransactionInstance-clone}{\code{TransactionInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransactionInstance-new"></a>}}
\if{latex}{\out{\hypertarget{method-TransactionInstance-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionInstance$new(transaction, config, endpoint = "insert")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transaction}}{(\code{list()})\cr
The transaction to be sent to the Fluree instance.}

\item{\code{config}}{(\code{list()})\cr
Configuration parameters of the instance.}

\item{\code{endpoint}}{(\code{string})\cr
The Fluree v4 endpoint to use: 'insert', 'upsert', or 'update'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransactionInstance-send"></a>}}
\if{latex}{\out{\hypertarget{method-TransactionInstance-send}{}}}
\subsection{Method \code{send()}}{
This method sends the configured transaction to the host.
The Fluree instance must be 'connected' before querying or transacting.
If \code{signMessages = TRUE} the JWT will be sent to the host.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionInstance$send()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransactionInstance-sign"></a>}}
\if{latex}{\out{\hypertarget{method-TransactionInstance-sign}{}}}
\subsection{Method \code{sign()}}{
This method is used to sign the configured transaction.
This method is called automatically when \code{signMessages = TRUE} and requires
a privateKey to be either passed as a parameter or configured within the
\code{config}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionInstance$sign(privateKey = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{privateKey}}{(\code{string})\cr
The private key to use for message signing (represented as a hex string).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{TransactionInstance}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransactionInstance-getSignedTransaction"></a>}}
\if{latex}{\out{\hypertarget{method-TransactionInstance-getSignedTransaction}{}}}
\subsection{Method \code{getSignedTransaction()}}{
Returns the signed transaction (if it has been set).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionInstance$getSignedTransaction()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{string}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransactionInstance-getTransaction"></a>}}
\if{latex}{\out{\hypertarget{method-TransactionInstance-getTransaction}{}}}
\subsection{Method \code{getTransaction()}}{
Returns the transaction body.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionInstance$getTransaction()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{string}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransactionInstance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransactionInstance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
