% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triples_to_jsonld.R
\name{triples_to_jsonld}
\alias{triples_to_jsonld}
\title{triples_to_jsonld}
\usage{
triples_to_jsonld(
  triples,
  context_df = NULL,
  base_uri = "https://example.org/survey/"
)
}
\arguments{
\item{triples}{data.frame containing RDF triples with columns \code{subject}, \code{predicate}, \code{object}, and \code{object_type}.
resulting from the \code{map_cto_to_rdf} function.}

\item{context_df}{data.frame containing the context for JSON-LD conversion. Should have columns \code{prefix} and \code{namespace}.}

\item{base_uri}{string representing the base URI for the JSON-LD document. Default is "https://example.org/survey/".}
}
\description{
Function to convert RDF triples to JSON-LD
}
\examples{
\dontrun{
srv <- "KiA_adaptation_ACTIVE"
kia_adapt <- novaCTO::readCTO(srv)
formdef <- kia_adapt$fromschema$kia_adaptation
fromRDF <- map_cto_to_rdf(formdef, base_uri = glue::glue("https://novapc.surveycto.com/{srv}/"), instrument = "KiA_adaptation_ACTIVE")
js <- triples_to_jsonld(fromRDF, make_surveycto_centext())
}

}
