% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configHandling.R
\name{updateConfig}
\alias{updateConfig}
\title{Update the Current Fluree Configuration}
\usage{
updateConfig(config, newConfig = list())
}
\arguments{
\item{config}{(\code{list()})\cr
The existing Fluree configuration parameters.}

\item{newConfig}{(\code{list()})\cr
The new parameters to configure the Fluree instance with.}
}
\value{
The new combined list of configuration parameters.
}
\description{
Update the configuration parameters of the current Fluree instance.
The \code{existingConfig} will be merged with the \code{newConfig} and all existing
fields will be replaced with the new ones if applicable; except for the
\code{defaultContext} field. Instead of replacing the context field the old and
new contexts are merged to form the updated \code{defaultContext}.
The merged \code{newConfig} is then returned.
}
\examples{
\dontrun{
newConfig <- list(ledger = "test2", port = "8090")
updatedConfig <- updateConfig(config, newConfig)

}
}
