% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchHandling.R
\name{validateBranchName}
\alias{validateBranchName}
\title{Validate a Fluree branch name}
\usage{
validateBranchName(branch)
}
\arguments{
\item{branch}{Branch name.}
}
\value{
\code{branch}, invisibly, when valid.
}
\description{
Validate a Fluree branch name
}
\keyword{internal}
