% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedGraphHandling.R
\name{validateGraphIri}
\alias{validateGraphIri}
\title{Validate a named-graph IRI}
\usage{
validateGraphIri(graph)
}
\arguments{
\item{graph}{Absolute graph IRI.}
}
\value{
\code{graph}, invisibly, when valid.
}
\description{
Validate a named-graph IRI
}
\keyword{internal}
