% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredicateInfo.R
\name{checkVNDomain}
\alias{checkVNDomain}
\title{Check: each property that has no domain has associated variable name}
\usage{
checkVNDomain(dom_res, ran_res, vn_res)
}
\arguments{
\item{dom_res}{\link{tibble} Containing columns: \code{predicate} and \code{domain}. The result of a SPARQL query to find the domains of all predicates in a list.}

\item{ran_res}{\link{tibble} Containing columns: \code{predicate} and \code{range}. The result of a SPARQL query to find the ranges of all predicates in a list.}

\item{vn_res}{\link{tibble} Containing two columns: \code{predicate} and \code{varname} (possibly NA).}
}
\value{
\link{tibble} Predicates for which no domain could be specified from \code{dom_res} and no variable name was supplied.
}
\description{
Helper function for \code{\link[=getPredicateInfo]{getPredicateInfo()}}.
}
