% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createArtistObject}
\alias{createArtistObject}
\title{createArtistObject}
\usage{
createArtistObject(name = NULL, entityId = NULL, deleteObject = FALSE)
}
\arguments{
\item{name}{character}
}
\value{
data.frame
}
\description{
function to create artist object
}
