% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createAuthObject}
\alias{createAuthObject}
\title{createAuthObject}
\usage{
createAuthObject(
  fuel = NULL,
  authority = NULL,
  type = NULL,
  doc = NULL,
  roles = NULL,
  salt = NULL,
  password = NULL,
  id = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{fuel}{character}

\item{authority}{character}

\item{type}{character}

\item{doc}{character}

\item{roles}{character}

\item{salt}{character}

\item{password}{character}

\item{id}{character}
}
\value{
data.frame
}
\description{
function to create auth object
}
