% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createAuthObject}
\alias{createAuthObject}
\title{createAuthObject}
\usage{
createAuthObject(id, doc, roles)
}
\arguments{
\item{id}{character}

\item{doc}{character}

\item{roles}{character}
}
\value{
data.frame
}
\description{
function to create auth object
}
