% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createCollectionObject}
\alias{createCollectionObject}
\title{createCollectionObject}
\usage{
createCollectionObject(name, doc, version)
}
\arguments{
\item{name}{character}

\item{doc}{character}

\item{version}{character}
}
\value{
data.frame
}
\description{
function to create collection object
}
