% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createFnObject}
\alias{createFnObject}
\title{createFnObject}
\usage{
createFnObject(
  language = NULL,
  spec = NULL,
  doc = NULL,
  code = NULL,
  params = NULL,
  name = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{language}{character}

\item{spec}{character}

\item{doc}{character}

\item{code}{character}

\item{params}{character}

\item{name}{character}
}
\value{
data.frame
}
\description{
function to create fn object
}
