% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlureeObjectFunctions.R
\name{createFnObject}
\alias{createFnObject}
\title{createFnObject}
\usage{
createFnObject(name, code, doc)
}
\arguments{
\item{name}{character}

\item{code}{character}

\item{doc}{character}
}
\value{
data.frame
}
\description{
function to create fn object
}
