% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLedger.R
\name{createLedger}
\alias{createLedger}
\title{createLedger}
\usage{
createLedger(
  ledgerName = "rdataset/mtcars",
  fluree_link = Sys.getenv("fluree_link")
)
}
\arguments{
\item{ledgerName}{Character. The name of the ledger to be created.}

\item{fluree_link}{Character. The link to the Fluree database.}
}
\value{
character
}
\description{
Helper function to create a ledger in Fluree
}
