% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledgerCreation.R
\name{createLedger}
\alias{createLedger}
\title{Create a new ledger}
\usage{
createLedger(ledgerName = NULL, config = NULL, transaction = NULL)
}
\arguments{
\item{ledgerName}{(\code{string})\cr
The name of the ledger to be created.}

\item{config}{(\code{list})\cr
The configuration parameters of the Fluree instance.}

\item{transaction}{(\code{list})\cr
The body of the initial transaction to add to the new ledger.}
}
\description{
This function creates a new ledger with the specified name. If a name is not
explicitly provided, the previously configured ledger name (stored in the
system environment) will be used. The same applies to the \code{config} parameters
of the Fluree instance. As for the \code{transaction} parameter if no
initial transaction is provided a default transaction will be sent to the newly
created ledger.
}
