% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlureeObjectFunctions.R
\name{createPredicateObject}
\alias{createPredicateObject}
\title{createPredicateObject}
\usage{
createPredicateObject(
  name,
  doc,
  type,
  restrictcollection,
  unique,
  multi,
  restricttag,
  index,
  upsert
)
}
\arguments{
\item{name}{character}

\item{doc}{character}

\item{type}{character}

\item{restrictcollection}{character}

\item{unique}{character}

\item{multi}{character}

\item{restricttag}{character}

\item{index}{character}

\item{upsert}{character}
}
\value{
data.frame
}
\description{
function to create predicate object
}
