% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createRoleObject}
\alias{createRoleObject}
\title{createRoleObject}
\usage{
createRoleObject(
  ctx = NULL,
  rules = NULL,
  doc = NULL,
  id = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{ctx}{character}

\item{rules}{character}

\item{doc}{character}

\item{id}{character}
}
\value{
data.frame
}
\description{
function to create role object
}
