% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlureeObjectFunctions.R
\name{createRuleObject}
\alias{createRuleObject}
\title{createRuleObject}
\usage{
createRuleObject(
  errormessage = NULL,
  collectiondefault = NULL,
  ops = NULL,
  fns = NULL,
  predicates = NULL,
  collection = NULL,
  doc = NULL,
  id = NULL,
  entityId = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{errormessage}{character}

\item{collectiondefault}{character}

\item{ops}{character}

\item{fns}{character}

\item{predicates}{character}

\item{collection}{character}

\item{doc}{character}

\item{id}{character}
}
\value{
data.frame
}
\description{
function to create rule object
}
