% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createSettingObject}
\alias{createSettingObject}
\title{createSettingObject}
\usage{
createSettingObject(
  language = NULL,
  id = NULL,
  txmax = NULL,
  passwords = NULL,
  doc = NULL,
  consensus = NULL,
  ledgers = NULL,
  anonymous = NULL,
  deleteObject = FALSE
)
}
\arguments{
\item{language}{character}

\item{id}{character}

\item{txmax}{character}

\item{passwords}{character}

\item{doc}{character}

\item{consensus}{character}

\item{ledgers}{character}

\item{anonymous}{character}
}
\value{
data.frame
}
\description{
function to create setting object
}
