% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlureeObjectFunctions.R
\name{createTagObject}
\alias{createTagObject}
\title{createTagObject}
\usage{
createTagObject(id, doc)
}
\arguments{
\item{id}{character}

\item{doc}{character}
}
\value{
data.frame
}
\description{
function to create tag object
}
