% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flureeObjectFunctions.R
\name{createUserObject}
\alias{createUserObject}
\title{createUserObject}
\usage{
createUserObject(username, auth, roles)
}
\arguments{
\item{username}{character}

\item{auth}{character}

\item{roles}{character}
}
\value{
data.frame
}
\description{
function to create user object
}
