% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{delete}
\alias{delete}
\title{Delete subjects by id}
\usage{
delete(id)
}
\arguments{
\item{id}{(\code{list()})\cr
The subject identifier/identifiers to retract from the Fluree instance.}
}
\description{
Delete is not an API endpoint in Fluree. This function merely transforms
a single or list of subject identifiers ( @id ) into a where/delete transaction
that deletes the subject and all facts about the subject.

Delete assumes that all facts for the provided subjects should be retracted
from the database.
}
