% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{generateFetchParams}
\alias{generateFetchParams}
\title{Constructs arguments for httr::POST}
\usage{
generateFetchParams(config, endpoint, contentType = "application/json")
}
\arguments{
\item{config}{(\code{list()})\cr
The configuration parameters of the current active transaction or query instance.}

\item{endpoint}{(\code{string})\cr
The endpoint to which the http request is made (either '/transact', '/query', '/create' or '/history').}

\item{contentType}{(\code{string})\cr
In the case of an unsigned message 'application/json' is used ('application/jwt' if signed).}
}
\value{
(\code{list()})
}
\description{
This function is a generic function to construct the parameters needed by
\code{httr} to execute a \code{POST()} request of a transaction or query.
}
