% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionUtils.R
\name{generateWhereDeleteForIds}
\alias{generateWhereDeleteForIds}
\title{Format a delete transaction}
\usage{
generateWhereDeleteForIds(ids, idAlias)
}
\arguments{
\item{idAlias}{(\code{string})\cr
The alias used for the @id field in the context of the Fluree instance.}

\item{id}{(\code{string})\cr
The entry to be deleted.}
}
\value{
(\code{list})
}
\description{
Generates the correct where-delete format to delete the given id.
}
