% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{getQuerySignature}
\alias{getQuerySignature}
\title{Get the Signed Query}
\usage{
getQuerySignature(queryVariables = NULL)
}
\arguments{
\item{queryVariables}{(\code{list()})\cr
The signed query instance to read the signature from.}
}
\value{
Character string representing the JWT of the signed query.
}
\description{
This function returns the JWT representation of the signed query.
Note this function can only be used if a private key had been configured and
the query has been signed.
}
\examples{
\dontrun{
queryInstance <- query(exampleQuery)
signedQueryInstance <- signQuery(queryInstance)

sig <- getQuerySignature(signedQueryInstance)
}

}
