% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{getQuerySignature}
\alias{getQuerySignature}
\title{Get the signed query}
\usage{
getQuerySignature()
}
\value{
String representation of the signed query as a JWT
}
\description{
This function returns the JWT representation of the signed query.
Note this function can only be used if a private key has been configured and
the query has already been signed.
}
