% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{getTransactionSignature}
\alias{getTransactionSignature}
\title{Get the signed transaction}
\usage{
getTransactionSignature()
}
\value{
String representation of the JWT of the signed transaction
}
\description{
This function returns the JWT representation of the signed transaction.
Note this function can only be used if a private key has been configured and
the transaction has already been signed.
}
