% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R, R/transactionHandling.R
\name{getTransactionText}
\alias{getTransactionText}
\title{Get the query}
\usage{
getTransactionText()

getTransactionText()
}
\value{
JSON string representation of the query body

JSON string representation of the transaction body
}
\description{
This function returns the query body as a JSON string.

This function returns the transaction body as a JSON string.
}
