% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{history}
\alias{history}
\title{History query configuration}
\usage{
history(query)
}
\arguments{
\item{query}{A list representing the body of the history query to be made}
}
\description{
This function configures the history query instance. System variables are set,
the default context is configured and if applicable the history query is signed.
}
