% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoryQueryInstance.R
\docType{class}
\name{HistoryQueryInstance}
\alias{HistoryQueryInstance}
\title{Class providing objects with methods to perform history queries on a Fluree instance.}
\description{
Class providing objects with methods to perform history queries on a Fluree instance.

Class providing objects with methods to perform history queries on a Fluree instance.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{query}}{(\code{list()})\cr
The list representation of a history query to be sent to the Fluree instance.}

\item{\code{config}}{(\code{list()})\cr
Configuration parameters of the instance.}

\item{\code{signedQuery}}{(\code{string})\cr
The JWT of the history query.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HistoryQueryInstance-new}{\code{HistoryQueryInstance$new()}}
\item \href{#method-HistoryQueryInstance-send}{\code{HistoryQueryInstance$send()}}
\item \href{#method-HistoryQueryInstance-sign}{\code{HistoryQueryInstance$sign()}}
\item \href{#method-HistoryQueryInstance-getSignedQuery}{\code{HistoryQueryInstance$getSignedQuery()}}
\item \href{#method-HistoryQueryInstance-getQuery}{\code{HistoryQueryInstance$getQuery()}}
\item \href{#method-HistoryQueryInstance-clone}{\code{HistoryQueryInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistoryQueryInstance-new"></a>}}
\if{latex}{\out{\hypertarget{method-HistoryQueryInstance-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistoryQueryInstance$new(query, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{(\code{list()})\cr
The history query to be sent to the Fluree instance.}

\item{\code{config}}{(\code{list()})\cr
Configuration parameters of the instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistoryQueryInstance-send"></a>}}
\if{latex}{\out{\hypertarget{method-HistoryQueryInstance-send}{}}}
\subsection{Method \code{send()}}{
This method sends the configured history query to the host.
The Fluree instance must be 'connected' before querying or transacting.
If \code{signMessages = TRUE} the JWT will be sent to the host.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistoryQueryInstance$send()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistoryQueryInstance-sign"></a>}}
\if{latex}{\out{\hypertarget{method-HistoryQueryInstance-sign}{}}}
\subsection{Method \code{sign()}}{
This method is used to sign the configured history query.
This method is called automatically when \code{signMessages = TRUE} and requires
a privateKey to be either passed as a parameter or configured within the
\code{config}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistoryQueryInstance$sign(privateKey = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{privateKey}}{(\code{string})\cr
The private key to use for message signing (represented as a hex string).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{HistoryQueryInstance}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistoryQueryInstance-getSignedQuery"></a>}}
\if{latex}{\out{\hypertarget{method-HistoryQueryInstance-getSignedQuery}{}}}
\subsection{Method \code{getSignedQuery()}}{
Returns the signed history query (if it has been set).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistoryQueryInstance$getSignedQuery()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{string}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistoryQueryInstance-getQuery"></a>}}
\if{latex}{\out{\hypertarget{method-HistoryQueryInstance-getQuery}{}}}
\subsection{Method \code{getQuery()}}{
Returns the history query body.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistoryQueryInstance$getQuery()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{string}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistoryQueryInstance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HistoryQueryInstance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistoryQueryInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
