% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R, R/Functions2.R
\name{insertData}
\alias{insertData}
\title{insertData}
\usage{
insertData(path, data_list, fluree_link = Sys.getenv("fluree_link"))

insertData(path, data_list, fluree_link = Sys.getenv("fluree_link"))
}
\arguments{
\item{path}{Character. The path to the Fluree database.}

\item{data_list}{List. The schema to be created (predicates).}

\item{fluree_link}{Character. The link to the Fluree database.}

\item{collectionName}{Character. The name of the Fluree database.}

\item{dataList}{Character. The list of data objects to add to the database.}
}
\value{
character

character
}
\description{
Helper function to helper create add predicates to a Fluree database.

Helper function to create a schema in Fluree
}
\examples{
: insertData("test", [{
 _id: '_predicate',
  name: '_auth/secretValue',
  type: 'string',
})

}
