% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_surveycto_context.R, R/reference_tables.R
\name{make_surveycto_centext}
\alias{make_surveycto_centext}
\title{make_surveycto_centext}
\usage{
make_surveycto_centext()

make_surveycto_centext()
}
\value{
tibble with columns \code{prefix} and \code{namespace}
}
\description{
Alternatiove duie to spelling error in original function name

This function creates a context for SurveyCTO RDF output. It defines the prefixes and namespaces used in the RDF triples.
}
\examples{
surveycto_centext <- make_surveycto_centext()
}
