% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextHandling.R
\name{mergeContexts}
\alias{mergeContexts}
\title{Merge Two Contexts}
\usage{
mergeContexts(context1, context2)
}
\arguments{
\item{context1}{The base context.\cr
Which can be a string, a list, or a named list.}

\item{context2}{The new context to merge,\cr
Which can also be a string, a list, or a named list.}
}
\value{
The merged context.\cr
If both contexts are strings, they are combined into a list.
}
\description{
Merges a new context into a base context
Supports strings, lists, and named lists.
Handles conflicts and duplicates in namespace abbreviations:
\itemize{
\item Solid duplicates (same key = same value) are kept once
\item Different abbreviations for the same namespace keep the first one
\item Same abbreviation for different namespaces: context1's value is kept
}
}
\examples{
mergeContexts("https://example.org/context1", "https://example.org/context2")
mergeContexts("https://example.org/context1", list("https://example.org/context2"))
mergeContexts(list("https://example.org/context1"), list("https://example.org/context2"))
mergeContexts(list("https://example.org/context1"), list(a = "https://example.org/context2"))

}
