% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextHandling.R
\name{mergeContexts}
\alias{mergeContexts}
\title{Merge Two Contexts}
\usage{
mergeContexts(context1, context2)
}
\arguments{
\item{context1}{The base context, which can be a string, a list, or a named list.}

\item{context2}{The new context to merge, which can also be a string, a list, or a named list.}
}
\value{
A merged context. If both contexts are strings, they are combined into a list.
}
\description{
Merges a new context into a base context, supporting strings, lists, and named lists (objects).
}
\examples{
merge_contexts("https://example.org/context1", "https://example.org/context2")
merge_contexts("https://example.org/context1", list("https://example.org/context2"))
merge_contexts(list("https://example.org/context1"), list("https://example.org/context2"))
merge_contexts(list("https://example.org/context1"), list(a = "https://example.org/context2"))
}
