% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer_with_type.R
\name{pivot_longer_with_type}
\alias{pivot_longer_with_type}
\title{pivot_longer_with_type}
\usage{
pivot_longer_with_type(data, names_to = "name", tripplenames = TRUE, ...)
}
\arguments{
\item{data}{A data frame to pivot longer. T}

\item{names_to}{Character. The name of the column to store the variable names in. Default is "name".}

\item{tripplenames}{Logical. If TRUE, the resulting column names will be "subject", "predicate", "object", and "type". If FALSE, the column names will be "ID", "name", "value", and "type".}

\item{...}{}
}
\description{
pivot_longer_with_type
}
