% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPredicates.R
\name{predicateTibble}
\alias{predicateTibble}
\title{Convert predicate mapping list to a tibble}
\usage{
predicateTibble(schema_list)
}
\arguments{
\item{schema_list}{\link{list} Predicate mapping list of lists generated by \code{\link[=mapPredicates]{mapPredicates()}}}
}
\value{
\link{tibble} Predicate mapping as a tibble. Columns: \code{predicate} (containing predicate IRI), \code{domain}, \code{range}, \code{varname} (corresponding column name)
}
\description{
Create predicate mapping tibble. Helper function for \code{\link[=mapPredicates]{mapPredicates()}} and \code{\link[=specIDPredicates]{specIDPredicates()}}
}
