% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacePrefixes.R
\name{prefixIRIs}
\alias{prefixIRIs}
\title{Replace IRIs with prefixes}
\usage{
prefixIRIs(data, add_context = NULL)
}
\arguments{
\item{data}{\link{tibble} Predicate mappings, most likely generated by \code{\link[=mapPredicates]{mapPredicates()}}}

\item{add_context}{\link{character} Additional prefix mappings as a named vector.}
}
\value{
\link{named list} Containing \code{data} = prefixed predicate mappings, \code{context} = context as a tibble
}
\description{
This function courtesy of ChatGPT.
This function replaces verbose IRIs in predicate mappings with their prefixed versions.
To do this, it uses the context list below, as well as additional prefix mappings provided in \code{add_context}
context <- c(
":" = "https://nova.org.za/nova-o#",
"aia" = "http://purl.org/aiaontology#",
"owl" = "http://www.w3.org/2002/07/owl#",
"rdf" = "http://www.w3.org/1999/02/22-rdf-syntax-ns#",
"xml" = "http://www.w3.org/XML/1998/namespace",
"xsd" = "http://www.w3.org/2001/XMLSchema#",
"foaf" = "http://xmlns.com/foaf/spec/#term_",
"prov" = "http://www.w3.org/ns/prov#",
"rdfs" = "http://www.w3.org/2000/01/rdf-schema#",
"time" = "https://www.w3.org/2006/time#",
"vcard" = "http://www.w3.org/TR/vcard-rdf/#",
"base" = "https://nova.org.za/nova-o"
)
}
\examples{
small_pred <- mapPredicates(varnames, predIRIs, domains, ranges)
geprefix <- prefixIRIs(small_pred)

}
