% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{query}
\alias{query}
\title{Query configuration}
\usage{
query(query)
}
\arguments{
\item{query}{A list representing the body of the query to be made}
}
\description{
This function configures the query instance. System variables are set,
the default context is configured and if applicable the query is signed.
}
