% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_from_tibble.R
\name{rdf_from_df}
\alias{rdf_from_df}
\title{rdf_from_df}
\usage{
rdf_from_df(
  df,
  subject = "subject",
  predicate = "predicate",
  object = "object",
  ...
)
}
\arguments{
\item{df}{Dataframe or tibble}

\item{subject}{Character. Default "subject"}

\item{predicate}{Character. Default "predicate"}

\item{object}{Character. Default "object"}
}
\value{
rdf object
}
\description{
rdf_from_df
}
\examples{
irisrdf <- rdf_from_df(iris)
}
