% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{sendHistoryQuery}
\alias{sendHistoryQuery}
\title{Send a history query}
\usage{
sendHistoryQuery(config)
}
\arguments{
\item{config}{A list of configuration parameters for the query}
}
\description{
This function makes use of httr to send the history query to the Fluree instance.
The response is then given as output.
}
