% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{sendHistoryQuery}
\alias{sendHistoryQuery}
\title{Send a history query}
\usage{
sendHistoryQuery(queryVariables)
}
\arguments{
\item{queryVariables}{(\code{list()})\cr
A list representing the history query specifications.}
}
\value{
A character string containing the response content.
}
\description{
This function makes use of \code{httr} to send the configured query to the
Fluree instance.
}
\examples{
historyQueryInstance <- history(exampleHistoryQuery)
sendHistoryQuery(historyQueryInstance)

}
