% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{sendQuery}
\alias{sendQuery}
\title{Send a query}
\usage{
sendQuery(config)
}
\arguments{
\item{config}{A list of configuration parameters for the query}
}
\description{
This function makes use of httr to send the query to the Fluree instance.
The response is then given as output.
}
