% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{sendQuery}
\alias{sendQuery}
\title{Send a query}
\usage{
sendQuery(queryVariables)
}
\arguments{
\item{queryVariables}{(\code{list()})\cr
A list representing the query specifications.}
}
\value{
A character string containing the response content.
}
\description{
This function makes use of \code{httr} to send the configured query to the
Fluree instance.
}
\examples{
queryInstance <- query(exampleQuery)
sendQuery(queryInstance)

}
