% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{sendTransaction}
\alias{sendTransaction}
\title{Send a Transaction}
\usage{
sendTransaction(transactionVariables)
}
\arguments{
\item{transactionVariables}{(\code{list()})\cr
A list representing the transaction specifications.}
}
\value{
A character string containing the response content.
}
\description{
This function makes use of \code{httr} package to send the configured transaction
to the Fluree \verb{/transact} API endpoint.
}
\examples{
transactionInstance <- transact(exampleData)
sendTransaction(transactionInstance)

}
