% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{sendTransaction}
\alias{sendTransaction}
\title{Send a transaction}
\usage{
sendTransaction(config)
}
\arguments{
\item{config}{A list of configuration parameters for the transaction}
}
\description{
This function makes use of httr to send the transaction to the Fluree instance.
The response is then given as output.
}
