% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configHandling.R
\name{setConfig}
\alias{setConfig}
\title{Set the Fluree Configuration Parameters}
\usage{
setConfig(host = NULL, port = NULL, ledger, signMessages = NULL)
}
\arguments{
\item{host}{(\code{string})\cr
The host name to send the transactions/queries to.}

\item{ledger}{(\code{string})\cr
The name of the ledger to be transacted to.}

\item{signMessages}{(\code{logical})\cr
Set to \code{TRUE} if all messages should be signed before transacting.}
}
\value{
The configuration parameters configured as a R list.
}
\description{
This function returns a default configuration for the Fluree instance.
It is invoked whenever a transaction/query is attempted without providing a
relevant config.
Optionally partial parameters may be provided with defaults configured for
any omitted ones.
}
\examples{
config <- setConfig(ledger = "test1")

}
