% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{signQuery}
\alias{signQuery}
\title{Sign a Query}
\usage{
signQuery(queryVariables = NULL, privateKey = NULL)
}
\arguments{
\item{queryVariables}{(\code{list()})\cr
A list representing the query to be signed.}

\item{privateKey}{(\code{character})\cr
The hexadecimal string representation of the private key to be used for signing.
If a private key is not explicitly provided, the one stored as an environment variable
will be used (if one had been configured).}
}
\value{
A list containing everything needed to query the Fluree database.
This includes all the necessary parameters as well as the signed query itself.
}
\description{
This function is used to sign a query which can then be sent to the Fluree
instance as a JWT (JSON Web Token).
}
\examples{
\dontrun{
queryInstance <- query(exampleQuery)
signedQueryInstance <- signQuery(queryInstance)
}

}
