% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHandling.R
\name{signQuery}
\alias{signQuery}
\title{Sign a query}
\usage{
signQuery(query = NULL, privateKey = NULL)
}
\arguments{
\item{query}{A list representing the query to be signed}

\item{privateKey}{The hexadecimal string representation of the private key to be used for signing}
}
\description{
This function is used to sign a query which can then be sent to the Fluree
instance as a JWT (JSON Web Token).
}
