% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{signTransaction}
\alias{signTransaction}
\title{Sign a transaction}
\usage{
signTransaction(transactionVariables = NULL, privateKey = NULL)
}
\arguments{
\item{transactionVariables}{(\code{list()})\cr
A list representing the transaction to be signed.}

\item{privateKey}{(\code{character})\cr
The hexadecimal string representation of the private key to be used for signing.
If a private key is not explicitly provided, the one stored as an environment variable
will be used (if one had been configured).}
}
\value{
A list containing everything needed to transact with Fluree.
This includes all the necessary parameters as well as the signed transaction itself.
}
\description{
This function is used to sign a transaction which can then be sent to the Fluree
instance as a JWT (JSON Web Token).
}
\examples{
\dontrun{
transactionInstance <- transact(exampleData)
signedTransactionInstance <- signTransaction(transactionInstance)

}
}
