% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{signTransaction}
\alias{signTransaction}
\title{Sign a transaction}
\usage{
signTransaction(transaction = NULL, privateKey = NULL)
}
\arguments{
\item{transaction}{A list representing the transaction to be signed}

\item{privateKey}{The hexadecimal string representation of the private key to be used for signing}
}
\description{
This function is used to sign a transaction which can then be sent to the Fluree
instance as a JWT (JSON Web Token).
}
