% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{transact}
\alias{transact}
\title{Transaction configuration}
\usage{
transact(
  config = NULL,
  ledger = NULL,
  transaction,
  signTransaction = NULL,
  privateKey = NULL,
  apiKey = NULL
)
}
\arguments{
\item{config}{(\code{list()})\cr
The configuration list for the Fluree instance.}

\item{ledger}{(\code{character})\cr
The name of the ledger to transact to.}

\item{transaction}{(\code{list()})\cr
The list representation of the transaction body to be sent.
Note alternatively the transaction can simply be passed as a JSON \code{character} string.}

\item{signTransaction}{(\code{logical})\cr
Determines whether the given transaction should be signed or not.}

\item{privateKey}{(\code{character})\cr
The hexstring representation of the private key to use for message signing.}
}
\value{
A list containing everything needed to transact with Fluree.
This includes all the necessary parameters as well as the signed/unsigned transaction itself.
}
\description{
This function configures the transaction as follows: the default context is configured
and if applicable the transaction is signed.
The function returns a list containing all the information necessary to interact
with the Fluree instance via \code{sendTransaction()}.
}
\examples{
\dontrun{
exampleData <- '{
   "insert": [
      {
        "@id": "freddy",
        "name": "Freddy"
      },
      {
        "@id": "alice",
        "name": "Alice"
      }
   ]
}'

transactionInstance <- transact(config = conf, ledger = 'demo', exampleData, signTransaction = FALSE)

# OR ALTERNATIVELY

dataList <- fromJSON(exampleData, simplifyDataFrame = FALSE, simplifyMatrix = FALSE, simplifyVector = FALSE)
transactionInstance <- transact(config = conf, ledger = 'demo', dataList, signTransaction = FALSE)

}
}
