% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionHandling.R
\name{transact}
\alias{transact}
\title{Transaction configuration}
\usage{
transact(transaction)
}
\arguments{
\item{transaction}{A list representing the body of the transaction to be sent}
}
\description{
This function configures the transaction as follows: system variables are set,
the default context is configured and if applicable the transaction is signed.
}
