% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startUp.R
\name{updateConfiguration}
\alias{updateConfiguration}
\title{Update the current Fluree configuration}
\usage{
updateConfiguration(newConfig = list())
}
\arguments{
\item{newConfig}{(\code{list()})\cr
The new parameters to configure the Fluree instance with.}
}
\description{
Update the configuration parameters of the current Fluree instance.
The existing \code{config} (stored in the system environment) will be merged with
the new one and all existing fields will be replaced with the new ones if
applicable; except for the \code{defaultContext} field. Instead of replacing the
\code{defaultContext} the old and new contexts are merged to form the updated \code{defaultContext}.
The new merged \code{config} will then be added back as an environment variable.
}
\examples{
newConfig <- list(ledger = "ledger2", port = "8090")
updateConfiguration(newConfig)

}
